Refactor: Split FlashTool class into mixins
