+++ lenz_flashtool/flashtool/core.py
@@ -12,39 +12,30 @@ This library provides functions for interfacing with BiSS C encoders using LENZ 
 and executing calibration routines.
 
 Author:
-    LENZ ENCODERS, 2020-2025
+    LENZ ENCODERS, 2020-2026
 '''
 
-import time
 import os
 import sys
 import logging
 import signal
-from typing import Callable, Dict, Optional, List, Union, Any, Type, Tuple, Literal
+from typing import Callable, Dict, Optional, Any, Tuple, Type
 from types import TracebackType
 import serial
 import serial.tools.list_ports
-import numpy as np
-from ..biss import (
-    biss_commands, interpret_biss_commandstate, interpret_error_flags, biss_crc6_calc,
-    BiSSBank,
-)
-from .uart import UartCmd, UartBootloaderCmd, UartBootloaderMemoryStates, UartBootloaderSeq
+
+from .serial_io import SerialIOMixin
+from .encoder_control import EncoderControlMixin
+from .biss_io import BiSSIOMixin
+from .data_streaming import DataStreamingMixin
+from .bootloader import BootloaderMixin
 from .errors import FlashToolError
-from .hex_utils import (
-    calculate_checksum,
-    generate_byte_line,
-    generate_hex_line,
-    reverse_endian, bytes_to_hex_str
-)
-from ..utils.progress import percent_complete
-from ..utils.termcolors import TermColors
-from .hex_utils import HexBlockExtractor
 
 logger = logging.getLogger(__name__)
 
 
-class FlashTool:
+class FlashTool(SerialIOMixin, BiSSIOMixin, EncoderControlMixin,
+                DataStreamingMixin, BootloaderMixin):
     """
     Main interface for interacting with BiSS C encoders via FlashTool device, connected to a serial port.
     FlashTool device has 2 channels for encoder connection. Commands perform on channel 2 if not mentioned.
@@ -116,21 +107,21 @@ class FlashTool:
         if hasattr(self, '_initialized') and self._initialized:
             return
         self._initialized = True
-        self.__port = None
-        self.__port_name = None
+        self._port = None
+        self._port_name = None
 
         ports = serial.tools.list_ports.comports(include_links=False)
         if os.name == 'nt':
             for porti in ports:
                 if porti.description.startswith(port_description_prefixes):
-                    self.__port_name = porti.device
+                    self._port_name = porti.device
                     try:
-                        self.__port = serial.Serial(porti.device, baud_rate, timeout=1)
-                        self.__port.set_buffer_size(rx_size=16777216, tx_size=16384)
-                        logger.debug('LENZ FlashTool %s - Connected!', self.__port_name)
+                        self._port = serial.Serial(porti.device, baud_rate, timeout=1)
+                        self._port.set_buffer_size(rx_size=16777216, tx_size=16384)
+                        logger.debug('LENZ FlashTool %s - Connected!', self._port_name)
                         break
                     except serial.SerialException as e:
-                        raise FlashToolError(f'LENZ FlashTool: {self.__port_name} is being used!') from e
+                        raise FlashToolError(f'LENZ FlashTool: {self._port_name} is being used!') from e
             else:
                 logger.error('Error: LENZ FlashTool not found!')
                 logger.debug('Program expectedly closed.')
@@ -138,27 +129,27 @@ class FlashTool:
         elif os.name == 'posix':
             found_port = self._find_linux_port_enhanced(ports, port_description_prefixes)
             if found_port:
-                self.__port_name = found_port
+                self._port_name = found_port
                 try:
-                    self.__port = serial.Serial(self.__port_name, baud_rate, timeout=1)
+                    self._port = serial.Serial(self._port_name, baud_rate, timeout=1)
 
-                    self.__port.bytesize = serial.EIGHTBITS
-                    self.__port.parity = serial.PARITY_NONE
-                    self.__port.stopbits = serial.STOPBITS_ONE
+                    self._port.bytesize = serial.EIGHTBITS
+                    self._port.parity = serial.PARITY_NONE
+                    self._port.stopbits = serial.STOPBITS_ONE
 
-                    self.__port.timeout = 1
-                    self.__port.write_timeout = 1
-                    self.__port.xonxoff = False
-                    self.__port.rtscts = False
+                    self._port.timeout = 1
+                    self._port.write_timeout = 1
+                    self._port.xonxoff = False
+                    self._port.rtscts = False
                     try:
-                        if hasattr(self.__port, 'set_buffer_sizes'):
+                        if hasattr(self._port, 'set_buffer_sizes'):
                             logger.debug("Has atribute")
-                            self.__port.set_buffer_sizes(rx_size=4096, tx_size=4096)
+                            self._port.set_buffer_sizes(rx_size=4096, tx_size=4096)
                     except Exception as e:
                         logger.debug(f"Buffer size adjustment not supported: {e}")
-                        logger.debug('LENZ FlashTool %s - Connected!', self.__port_name)
+                        logger.debug('LENZ FlashTool %s - Connected!', self._port_name)
                 except serial.SerialException as e:
-                    raise FlashToolError(f'LENZ FlashTool: {self.__port_name} is being used!') from e
+                    raise FlashToolError(f'LENZ FlashTool: {self._port_name} is being used!') from e
             else:
                 logger.error('Error: LENZ FlashTool not found!')
                 logger.debug('Available ports: %s', [f"{p.device}: {p.description} (HWID: {p.hwid})" for p in ports])
@@ -166,55 +157,7 @@ class FlashTool:
         else:
             logger.error('Error: Unsupported operating system!')
             raise FlashToolError('Unsupported operating system!')
-        self.__port.flushInput()
-
-    def _find_linux_port_enhanced(self, ports, port_description_prefixes):
-        """
-        Enhanced Linux port detection for FlashTool devices.
-
-        Searches for available serial ports that match the specified criteria,
-        with priority given to XR21V1410 devices.
-
-        Args:
-            ports: List of serial port objects obtained from serial.tools.list_ports.comports()
-            port_description_prefixes: Tuple of string prefixes to match against port descriptions
-                                    (e.g., ('XR21V',) for XR21V1410 devices)
-
-        Returns:
-            str or None: Device path (e.g., '/dev/ttyUSB0') if a matching port is found,
-                        None if no compatible device is detected.
-
-        Search Patterns (in order of priority):
-            1. VID:PID matching - Looks for '04e2:1410' in hardware ID (most reliable)
-            2. Description matching - Checks if port description contains any of the prefixes
-            3. Manufacturer matching - Checks if manufacturer field contains any of the prefixes  
-            4. Product matching - Checks if product field contains any of the prefixes
-
-        Example:
-            >>> ports = serial.tools.list_ports.comports()
-            >>> device = self._find_linux_port_enhanced(ports, ('XR21V',))
-            >>> print(device)
-            '/dev/ttyUSB0'
-        """
-
-        search_patterns = [
-            # VID:PID (04e2:1410 for XR21V1410)
-            lambda p: '04e2:1410' in p.hwid.lower(),
-            lambda p: p.description and any(prefix in p.description for prefix in port_description_prefixes),
-            lambda p: p.manufacturer and any(prefix in p.manufacturer for prefix in port_description_prefixes),
-            lambda p: p.product and any(prefix in p.product for prefix in port_description_prefixes),
-        ]
-
-        for port in ports:
-            for pattern in search_patterns:
-                try:
-                    if pattern(port):
-                        logger.debug(f"Found port: {port.device} - {port.description} (HWID: {port.hwid})")
-                        return port.device
-                except Exception as e:
-                    logger.debug(f"Error checking pattern on port {port.device}: {e}")
-                    continue
-        return None
+        self._port.flushInput()
 
     def __enter__(self) -> 'FlashTool':
         """
@@ -336,66 +279,12 @@ class FlashTool:
             >>> ft = FlashTool()
             >>> ft._default_cleanup()  # Closes the serial port
         """
-        if self.__port is not None and self.__port.is_open:
+        if self._port is not None and self._port.is_open:
             try:
-                self.__port.close()
-                logger.debug("Serial port %s closed", self.__port_name)
+                self._port.close()
+                logger.debug("Serial port %s closed", self._port_name)
             except Exception as e:
                 logger.warning("Error closing port: %s", e)
-
-    def _wait_for_data(self, size: int, timeout: float = 1.0) -> bool:
-        """
-        Waits for the specified amount of data to be available in the serial input buffer.
-
-        Polls the serial port until the required number of bytes is available
-        or the timeout expires, checking every 10 milliseconds.
-
-        Args:
-            size: The number of bytes to wait for.
-            timeout: Maximum time to wait in seconds. Defaults to 1.0.
-
-        Returns:
-            bool: True if the required data is available before the timeout, False otherwise.
-
-        Example:
-            >>> ft = FlashTool()
-            >>> ft._wait_for_data(10, timeout=2.0)
-            True
-        """
-        start_time = time.time()
-        while (time.time() - start_time) < timeout:
-            if self.__port.in_waiting >= size:
-                return True
-            time.sleep(0.01)
-        return False
-
-    def _write_to_port(self, data: bytes) -> None:
-        """
-        Writes data to the serial port with unified error handling.
-
-        Handles all serial port write operations, logging errors and raising exceptions for communication failures.
-
-        Args:
-            data: The bytes to write to the serial port.
-
-        Raises:
-            FlashToolError: If a serial communication error or unexpected failure occurs.
-
-        Example:
-            >>> ft = FlashTool()
-            >>> ft._write_to_port(b'\\x01\\x00@\\x0b\\xff\\xb5')  # Power off encoder on channel 2
-        """
-        try:
-            self.__port.reset_input_buffer()
-            self.__port.reset_output_buffer()
-            self.__port.write(data)
-            self.__port.flush()
-        except (serial.SerialException, OSError) as e:
-            logger.error("Port write failed: %s", e)
-            raise FlashToolError(f"Hardware communication failed: {e}") from e
-        except Exception as e:
-            logger.critical("Unexpected port write error: %s", e)
-            raise FlashToolError("Critical write operation failure") from e
 
     def close(self):
         """
@@ -431,2313 +320,4 @@ class FlashTool:
         # Reset singleton instance
         FlashTool._instance = None
         self._initialized = False
-        logger.info('FlashTool: %s closed.', self.__port_name)
-
-    def port_read(self, length: int) -> np.ndarray:
-        """
-        Reads a BiSS frame from the serial port, validates its checksum, and returns the payload.
-
-        The method blocks until the expected number of bytes arrives (or a timeout occurs).
-        A complete BiSS frame consists of:
-
-        * 1 byte  - length
-        * 2 bytes - address (repeated)
-        * 1 byte  - command
-        * ``length`` bytes - payload data
-        * 1 byte  - checksum
-
-        The checksum is verified with :func:`calculate_checksum` over **all bytes except the
-        checksum itself**.  If verification succeeds the payload (the ``length`` data bytes)
-        is returned as a ``uint8`` NumPy array.  On failure a ``FlashToolError`` is raised.
-
-        Args:
-            length: Number of **payload** bytes expected (excludes the 5 technical bytes:
-                    length, address*2, command, checksum).
-
-        Returns:
-            np.ndarray: 1-D array of ``dtype=uint8`` containing only the payload data.
-
-        Raises:
-            FlashToolError:
-                * If no data arrives within the internal timeout (currently 1 s).
-                * If the received checksum does not match the calculated one.
-
-        Example:
-            >>> ft = FlashTool()
-            >>> data = ft.port_read(10)
-            >>> print(data)
-            array([0x01, 0x02, ..., 0x0A], dtype=uint8)
-        """
-        if self._wait_for_data(length + 1, timeout=1.0):
-            biss_data = self.__port.read(length + 5)  # len, addr, addr, cmd, checksum
-            biss_value = int.from_bytes(biss_data, byteorder='big', signed=False)
-            calculated_crc = calculate_checksum(biss_data[0:-1].hex())
-            if calculated_crc == biss_data[-1]:
-                crc_res = "OK"
-                logger.debug(f"Received BiSS Data: {biss_value:#010x}, checksum calculated {calculated_crc}, \
-                            in data {biss_data[-1]}, res = {crc_res}")
-            else:
-                crc_res = "FALSE"
-                logger.error(f"Received BiSS Data: {biss_value:#010x}, checksum calculated {calculated_crc}, \
-                            in data {biss_data[-1]}, res = {crc_res}")
-            logger.debug("BiSS received data:")
-            data_array = np.array(list(biss_data[4:-1]), 'uint8')
-            logger.debug(data_array)
-            return data_array
-            # logger.info(np.array(list(biss_data[0:64]), 'uint8'))
-            # logger.info(np.array(list(biss_data[64:-1]), 'uint8'))
-        raise FlashToolError('Timeout waiting for register data.')
-
-    def biss_cmd_reboot2bl(self):
-        """
-        Sends the BiSS command to reboot the device into bootloader mode.
-
-        Issues the 'reboot2bl' command to the BiSS encoder, initiating a reboot into bootloader
-        mode for firmware updates. Note: This method is deprecated and should be replaced with `biss_write_command('reboot2bl')`.
-
-        Returns:
-            None
-
-        Example:
-            >>> ft = FlashTool()
-            >>> ft.biss_cmd_reboot2bl()
-            WARNING: Deprecated method. Use biss_write_command('reboot2bl') instead.
-        """
-        logger.debug("Sending reboot to bootloader command")
-        command_code: int = biss_commands['reboot2bl'][0]
-        command_list: list[int] = [command_code & 0xFF, (command_code >> 8) & 0xFF]
-        self.biss_write_word(BiSSBank.CMD_REG_INDEX, command_list)
-        time.sleep(0.5)
-
-    def biss_write_command(self, command: str):
-        """
-        Sends a specific BiSS command to the encoder.
-
-        Issues a command from the `biss_commands` dictionary to the BiSS encoder, such as rebooting,
-        setting direction, or initiating calibration. Commands are written to the command register.
-
-        Args:
-            command: The command key from `biss_commands`. Supported commands include:
-
-                - 'non': No operation command. Device remains in current state with no action taken.
-                - 'load2k': Load 2 KB of data from registers into device memory. Used for firmware updates.
-                - 'staybl': Force device to stay in bootloader mode.
-                - 'run': Exit bootloader mode and execute the main application program. Initiates normal encoder operation.
-                - 'zeroing': Perform encoder zeroing procedure. Resets position counter to zero at current mechanical position.
-                - 'set_dir_cw': Configure encoder direction sensing for clockwise rotation.
-                - 'set_dir_ccw': Configure encoder direction sensing for counterclockwise rotation.
-                - 'saveflash': Commit current configuration parameters to flash memory.
-                - 'ampcalibrate': Initiate amplitude calibration routine. Automatically adjusts signal amplitudes.
-                - 'cleardiflut': Clear differential lookup table (DifLUT). Resets all offset compensation values to default state.
-                - 'set_fast_biss': Enable high-speed BiSS communication mode.
-                - 'set_default_biss': Revert to standard BiSS communication mode.
-                - 'reboot': Perform device reboot.
-                - 'reboot2bl': Reboot device directly into bootloader mode.
-                - 'loadseriald': Write serial number and manufacturing date to OTP memory.
-                - 'loaddevid': Program DevID into OTP memory.
-                - 'loadkey': Store key in secure flash memory.
-                - 'savediftable': Save differential compensation table to memory.
-                - 'unlockflash': Enable write access to protected flash memory regions.
-                - 'unlocksetup': Disable configuration write protection.
-                - 'enarccal': Enable arc-based calibration mode.
-                - 'clearcalflash': Erase all high-speed oscillator (HSI) and differential calibration data from flash.
-
-        Raises:
-            FlashToolError: If the specified command is not found in `biss_commands`.
-
-        Example:
-            >>> ft = FlashTool()
-            >>> ft.biss_write_command('reboot2bl')
-        """
-        if command in biss_commands:
-            command_code: int = biss_commands[command][0]
-            command_list: list[int] = [command_code & 0xFF, (command_code >> 8) & 0xFF]
-            self._write_to_port(generate_byte_line(BiSSBank.CMD_REG_INDEX, UartCmd.HEX_WRITE_CMD, command_list))
-            logger.debug("Sending BiSS %s command (%s)", command, hex(command_code))
-            time.sleep(0.01)
-        else:
-            logger.error("Unknown command: '%s'", command)
-            raise logger(f"Unknown command: '{command}'.")
-
-    def hex_line_send(self, hex_line: str) -> bytes:
-        """
-        Sends a hex-formatted line to the FlashTool device.
-
-        Converts the provided hex line (starting with ':') to bytes and transmits it to the FlashTool device via the serial port.
-
-        Args:
-            hex_line: A string representing a hex line, including the leading ':'.
-
-        Returns:
-            bytes: The transmitted bytes.
-
-        Example:
-            >>> ft = FlashTool()
-            >>> ft.hex_line_send(':0100400C0AA9')
-            b'\\x01\\x00@\\x0c\\n\\xa9'
-        """
-        # logger.debug(f'Uploading {hex_line} to the FlashTool.')
-        tx_row = bytes.fromhex(hex_line[1:])
-        # logger.debug(tx_row.hex())
-        self._write_to_port(tx_row)
-        return tx_row
-
-    def biss_set_bank(self, bank_num: int) -> None:
-        """
-        Sets the active BiSS bank for subsequent read/write operations.
-
-        Selects the specified bank number for BiSS register operations, ensuring it is within the valid range (0-255).
-
-        Args:
-            bank_num: The BiSS bank number to select.
-
-        Raises:
-            ValueError: If the bank number is out of the valid range (0-255).
-            TypeError: If the bank number is not an integer.
-
-        Example:
-            >>> ft = FlashTool()
-            >>> ft.biss_set_bank(1)
-        """
-        try:
-            if not 0 <= bank_num <= 255:
-                raise ValueError(f"Bank {bank_num} out of range (0-255)")
-            logger.debug('Setting bank %s', bank_num)
-            self._write_to_port(generate_byte_line(BiSSBank.BSEL_REG_INDEX, UartCmd.HEX_WRITE_CMD, [bank_num]))
-        except (ValueError, TypeError) as e:
-            logger.error("Validation error in biss_set_bank: %s", str(e))
-            raise
-
-    def biss_write(self, addr: int, data: int) -> None:
-        """
-        Writes a single byte to a BiSS register at the specified address.
-
-        Sends a single 8-bit value to the specified BiSS register address.
-        Note: This method is deprecated and may be removed in future versions.
-
-        Args:
-            addr: The BiSS register address (0-127).
-            data: The 8-bit integer value to write.
-
-        Raises:
-            ValueError: If the address is out of range (0-127) or if the data is negative.
-            TypeError: If the data is not an integer.
-
-        Example:
-            >>> ft = FlashTool()
-            >>> ft.biss_write(0x10, 0xFF)
-            WARNING: Deprecated method. Consider using biss_write_word instead.
-        """
-        # TODO unused function?
-        try:
-            if not 0 <= addr <= 127:
-                raise ValueError(f"Address {addr} out of range (0-127)")
-
-            if not isinstance(data, int):
-                raise TypeError(f"Data {data} is not an integer (got {type(data)})")
-            if data < 0:
-                raise ValueError(f"Data {data} is negative")
-            self._write_to_port(generate_byte_line(addr, UartCmd.HEX_WRITE_CMD, [data]))
-        except (ValueError, TypeError) as e:
-            logger.error("Validation error in biss_write: %s", str(e))
-            raise
-
-    def biss_write_word(self, addr: int, word: Union[int, List[int], bytearray, np.ndarray]) -> None:
-        """
-        Writes one or more 8, 16, or 32-bit words to BiSS registers starting at the specified address.
-
-        Converts the provided word(s) to bytes, handles endianness, and writes them to the BiSS encoder.
-        Supports:
-        - Single integers or lists of integers
-        - bytearray
-        - numpy arrays (np.int8, np.int16, np.int32, np.uint8, np.uint16, np.uint32)
-
-        Args:
-            addr: The starting BiSS register address (0-127).
-            word: A single integer, list of integers, bytearray, or numpy array to write.
-
-        Raises:
-            ValueError: If the address is out of range, the word list is empty, a word is negative,
-            or a word exceeds the 32-bit limit.
-            TypeError: If any word is not an integer.
-            FlashToolError: If a hardware communication error occurs.
-
-        Example:
-            >>> ft = FlashTool()
-            >>> ft.biss_write_word(0x10, 0xABCD)  # Write a 16-bit word
-            >>> ft.biss_write_word(0x20, [0x01, 0x02])  # Write two 8-bit words
-            >>> ft.biss_write_word(0x30, np.array([0x1234, 0x5678], dtype=np.int16))
-            >>> ft.biss_write_word(0x40, bytearray([0x01, 0x02, 0x03]))
-        """
-        try:
-            if not 0 <= addr <= 127:
-                raise ValueError(f"Address {addr} out of range (0-127)")
-
-            if isinstance(word, (np.int8, np.int16, np.int32, np.uint8, np.uint16, np.uint32)):
-                words = [int(word)]
-                word_size = word.itemsize
-
-            elif isinstance(word, int):
-                words = [word]
-                max_word = word
-                if max_word > 0xFFFFFFFF:
-                    raise ValueError(f"Word value {max_word} exceeds 32-bit limit")
-                word_size = 4 if max_word > 0xFFFF else (2 if max_word > 0xFF else 1)
-
-            elif isinstance(word, np.ndarray):
-                if word.ndim != 1:
-                    raise ValueError(f"Numpy array must be 1-dimensional (got {word.ndim}D)")
-
-                supported_dtypes = [
-                    np.int8, np.int16, np.int32, np.uint8, np.uint16, np.uint32,
-                    'int8', 'int16', 'int32', 'uint8', 'uint16', 'uint32'
-                ]
-
-                if word.dtype not in supported_dtypes:
-                    raise ValueError(f"Unsupported numpy dtype: {word.dtype}. "
-                                     f"Supported: int8, int16, int32, uint8, uint16, uint32")
-
-                words = word.tolist()
-                word_size = word.itemsize
-
-                if word.dtype in [np.uint8, 'uint8'] and any(w > 0xFF for w in words):
-                    raise ValueError("uint8 values must be <= 0xFF")
-                elif word.dtype in [np.uint16, 'uint16'] and any(w > 0xFFFF for w in words):
-                    raise ValueError("uint16 values must be <= 0xFFFF")
-                elif word.dtype in [np.uint32, 'uint32'] and any(w > 0xFFFFFFFF for w in words):
-                    raise ValueError("uint32 values must be <= 0xFFFFFFFF")
-
-            elif isinstance(word, bytearray):
-                words = list(word)
-                word_size = 1
-
-            elif isinstance(word, list):
-                words = word
-                if not words:
-                    raise ValueError("Empty word list provided")
-
-                for i, w in enumerate(words):
-                    if not isinstance(w, int):
-                        raise TypeError(f"Word {i} is not an integer (got {type(w)})")
-                    if w < 0:
-                        raise ValueError(f"Word {i} is negative ({w})")
-
-                max_word = max(words)
-                if max_word > 0xFFFFFFFF:
-                    raise ValueError(f"Word value {max_word} exceeds 32-bit limit")
-                word_size = 4 if max_word > 0xFFFF else (2 if max_word > 0xFF else 1)
-
-            else:
-                raise TypeError(f"Unsupported type: {type(word)}. "
-                                f"Supported: int, list[int], bytearray, np.ndarray")
-
-            if not isinstance(word, (np.ndarray, np.integer)):
-                max_val_in_list = max(words) if words else 0
-                if word_size == 1 and max_val_in_list > 0xFF:
-                    raise ValueError(f"Word size determined as 1-byte, but value {max_val_in_list} > 0xFF")
-                elif word_size == 2 and max_val_in_list > 0xFFFF:
-                    raise ValueError(f"Word size determined as 2-byte, but value {max_val_in_list} > 0xFFFF")
-                elif word_size == 4 and max_val_in_list > 0xFFFFFFFF:
-                    raise ValueError(f"Word size determined as 4-byte, but value {max_val_in_list} > 0xFFFFFFFF")
-
-            byte_list = []
-
-            if isinstance(word, np.ndarray):
-                byte_list = list(word.tobytes())
-
-            elif isinstance(word, bytearray):
-                byte_list = list(word)
-
-            else:
-                for w in words:
-                    try:
-                        byte_list.extend(w.to_bytes(word_size, byteorder='big', signed=(w < 0)))
-                    except OverflowError as e:
-                        raise ValueError(f"Word {w} doesn't fit in {word_size} byte(s)") from e
-
-            reversed_bytes = reverse_endian(byte_list, word_size)
-
-            if isinstance(word, np.ndarray):
-                logger.debug("Sending numpy array %s (dtype: %s, shape: %s) to address %s",
-                             word, word.dtype, word.shape, addr)
-            else:
-                logger.debug("Sending word %s with starting index %s", word, addr)
-
-            self._write_to_port(generate_byte_line(addr, UartCmd.HEX_WRITE_CMD, reversed_bytes))
-
-        except (ValueError, TypeError) as e:
-            logger.error("Validation error in biss_write_word: %s", str(e))
-            raise
-
-    def biss_read_state_flags(self) -> np.ndarray:
-        """
-        Reads the state flags from the BiSS encoder.
-
-        Sends a read command to retrieve the state flags, validates the response
-        with a checksum check, and returns the flags as a NumPy array.
-
-        Returns:
-            np.ndarray: A NumPy array of uint8 values containing the state flags.
-
-        Raises:
-            FlashToolError: If the checksum validation fails or no data is received within the timeout.
-
-        Example:
-            >>> ft = FlashTool()
-            >>> flags = ft.biss_read_state_flags()
-            >>> print(flags)
-            array([0x01, 0x00], dtype=uint8)
-        """
-        self._write_to_port(generate_byte_line(BiSSBank.STATE_FLAG_REG_INDEX, UartCmd.HEX_READ_CMD, [1, 2]))
-        self.__port.reset_input_buffer()
-        time.sleep(0.01)
-
-        if self.__port.in_waiting >= 1:
-            biss_data = self.__port.read(7)
-            biss_value = int.from_bytes(biss_data, byteorder='big', signed=False)
-            if (calculate_checksum(biss_data[0:-1].hex())) == (biss_data[-1]):
-                crc_res = "OK"
-            else:
-                crc_res = "FALSE"
-                logger.error(f"Received BiSS Data: {biss_value:#010x}, \
-                              checksum calculated {calculate_checksum(biss_data[0:-1].hex())}, \
-                              in data {(biss_data[-1])}, \tres = {crc_res}")
-                raise FlashToolError('checksum validation failed for state flags.')
-            logger.debug("Received BiSS Data: %s", np.array(list(biss_data[:-1]), 'uint8'))
-            logger.debug(f"Received BiSS Data: {biss_value:#010x}, \
-                          checksum calculated {calculate_checksum(biss_data[0:-1].hex())}, \
-                          in data {(biss_data[-1])}, \tres = {crc_res}")
-            return (np.array(list(biss_data[4:-1]), 'uint8'))
-        raise FlashToolError('No answer from encoder.')
-
-    def biss_read_registers(self, bissbank: int) -> None:
-        """
-        Reads all registers from the specified BiSS bank and logs the results.
-
-        Selects the specified bank, sends a read command for all registers,
-        validates the response with a checksum check, and logs the received data as NumPy arrays.
-        Does not return the data, only logs it for debugging purposes.
-
-        Args:
-            bissbank: The BiSS bank number to read from (0-255).
-
-        Raises:
-            FlashToolError: If the read operation times out or a communication error occurs.
-
-        Example:
-            >>> ft = FlashTool()
-            >>> ft.biss_read_registers(1)
-            INFO: BiSS registers:
-            ...
-        """
-        self.biss_set_bank(bissbank)
-        # self.__port.flushInput()
-        self.__port.reset_input_buffer()
-        self._write_to_port(
-                            generate_byte_line(0, UartCmd.HEX_READ_CMD,
-                                               [item for item in list(range(BiSSBank.REGISTER_PLUS_FIXED_BANK_SIZE))]))
-        time.sleep(0.1)
-
-        if self.__port.in_waiting >= 4:
-            biss_data = self.__port.read(BiSSBank.REGISTER_PLUS_FIXED_BANK_SIZE + 5)  # len, addr, addr, cmd, crc
-            biss_value = int.from_bytes(biss_data, byteorder='big', signed=False)
-            if (calculate_checksum(biss_data[0:-1].hex())) == (biss_data[-1]):
-                crc_res = "OK"
-            else:
-                crc_res = "FALSE"
-                logger.error(f"Received BiSS Data: {biss_value:#010x}, \
-                            checksum calculated {calculate_checksum(biss_data[0:-1].hex())}, \
-                            in data {(biss_data[-1])}, \tres = {crc_res}")
-            logger.info("BiSS registers:")
-            logger.info(np.array(list(biss_data[4:68]), 'uint8'))
-            logger.info(np.array(list(biss_data[68:-1]), 'uint8'))
-
-    def encoder_power_off(self) -> None:
-        """
-        Powers off the encoder on channel 2.
-
-        Sends a POWER_OFF command to the BiSS encoder connected to channel 2 of the FlashTool device.
-
-        Returns:
-            None
-
-        Example:
-            >>> ft = FlashTool()
-            >>> ft.encoder_power_off()
-        """
-        logger.debug('Sending POWER_OFF command to the encoder')
-        tx_row = bytes.fromhex(generate_hex_line(0, UartCmd.CMD_POWER_OFF, [0])[1:])
-        # logger.debug(tx_row)
-        self._write_to_port(tx_row)
-
-    def encoder_power_on(self) -> None:
-        """
-        Powers on the encoder on channel 2.
-
-        Sends a POWER_ON command to the BiSS encoder connected to channel 2 of the FlashTool device.
-
-        Returns:
-            None
-
-        Example:
-            >>> ft = FlashTool()
-            >>> ft.encoder_power_on()
-        """
-        logger.debug('Sending POWER_ON command to the encoder')
-        tx_row = bytes.fromhex(generate_hex_line(0, UartCmd.CMD_POWER_ON, [0])[1:])
-        # logger.debug(tx_row)
-        self._write_to_port(tx_row)
-
-    def encoder_ch1_power_off(self) -> None:
-        """
-        Powers off the encoder on channel 1.
-
-        Sends a POWER_OFF command to the BiSS encoder connected to channel 1 of the FlashTool device.
-
-        Returns:
-            None
-
-        Example:
-            >>> ft = FlashTool()
-            >>> ft.encoder_ch1_power_off()
-        """
-        logger.debug('Sending POWER_OFF command to the encoder')
-        tx_row = bytes.fromhex(generate_hex_line(0, UartCmd.CMD_CH1_POWER_OFF, [0])[1:])
-        # logger.debug(tx_row)
-        self._write_to_port(tx_row)
-
-    def encoder_ch1_power_on(self) -> None:
-        """
-        Powers on the encoder on channel 1.
-
-        Sends a POWER_ON command to the BiSS encoder connected to channel 1 of the FlashTool device.
-
-        Returns:
-            None
-
-        Example:
-            >>> ft = FlashTool()
-            >>> ft.encoder_ch1_power_on()
-        """
-        logger.debug('Sending POWER_ON command to the encoder')
-        tx_row = bytes.fromhex(generate_hex_line(0, UartCmd.CMD_CH1_POWER_ON, [0])[1:])
-        logger.debug(tx_row.hex())
-        self._write_to_port(tx_row)
-
-    def encoder_power_cycle(self) -> None:
-        """
-        Performs a power cycle on the encoder on channel 2.
-
-        Powers off the encoder, waits 0.1 seconds, powers it back on, and waits another 0.1 seconds to ensure stabilization.
-
-        Returns:
-            None
-
-        Example:
-            >>> ft = FlashTool()
-            >>> ft.encoder_power_cycle()
-        """
-        self.encoder_power_off()
-        time.sleep(0.1)
-        self.encoder_power_on()
-        time.sleep(0.1)
-        logger.debug('Performed power cycle to the encoder.')
-
-    def encoder_ch1_power_cycle(self) -> None:
-        """
-        Performs a power cycle on the encoder on channel 1.
-
-        Powers off the encoder, waits 0.1 seconds, powers it back on, and waits another 0.1 seconds to ensure stabilization.
-
-        Returns:
-            None
-
-        Example:
-            >>> ft = FlashTool()
-            >>> ft.encoder_ch1_power_cycle()
-        """
-        self.encoder_ch1_power_off()
-        time.sleep(0.1)
-        self.encoder_ch1_power_on()
-        time.sleep(0.1)
-        logger.debug('Performed power cycle to the encoder.')
-
-    def flashtool_rst(self) -> None:
-        """
-        Resets the FlashTool device.
-
-        Sends a RESET command to the FlashTool, reinitializing its internal state.
-
-        Returns:
-            None
-
-        Example:
-            >>> ft = FlashTool()
-            >>> ft.flashtool_rst()
-        """
-        logger.info('Sending RESET command to FlashTool')
-        tx_row = bytes.fromhex(generate_hex_line(0, UartCmd.CMD_NVRST, [0])[1:])
-        # logger.debug(tx_row)
-        self._write_to_port(tx_row)
-        time.sleep(0.01)
-
-    def reboot_to_bl(self) -> None:
-        """
-        Reboot to BOOTLOADER FlashTool device.
-
-        Sends a Reboot to Bootloader command to the FlashTool.
-
-        Returns:
-            None
-
-        Example:
-            >>> ft = FlashTool()
-            >>> ft.reboot_to_bl()
-        """
-        logger.info('Sending REBOOT to BOOTLOADER command to FlashTool')
-        tx_row = bytes.fromhex(generate_hex_line(
-            address=0x0000,
-            command=UartCmd.CMD_REBOOT_TO_BL,
-            data=[0x00]
-        )[1:])
-        self._write_to_port(tx_row)
-        time.sleep(0.01)
-
-    def reboot_to_fw(self):
-        """
-        Reboot to FIRMWARE FlashTool device.
-
-        Sends a Reboot to Firmware command to the FlashTool.
-
-        Returns:
-            None
-
-        Example:
-            >>> ft = FlashTool()
-            >>> ft.reboot_to_fw()
-        """
-        logger.info('Sending REBOOT to FIRMWARE command to FlashTool')
-        tx_row = bytes.fromhex(generate_hex_line(
-            address=0x0000,
-            command=UartBootloaderCmd.UART_COMMAND_RUN_PROGRAM,
-            data=[0x00]
-        )[1:])
-        logger.debug(f"Sent BiSS Data: {tx_row.hex()}")
-        self._write_to_port(tx_row)
-        time.sleep(0.01)
-
-    def read_fw_bl_ver(self) -> tuple[str, str]:
-        """
-        Read firmware and bootloader version from FlashTool device.
-
-        Sends a command to read both firmware and bootloader versions from the device.
-        The response contains 8 bytes where:
-
-        * First 4 bytes represent firmware version
-        * Last 4 bytes represent bootloader version
-
-        Returns:
-            tuple[str, str]: A tuple containing:
-                - Firmware version as hexadecimal string (4 characters)
-                - Bootloader version as hexadecimal string (4 characters)
-
-        Example:
-            >>> ft = FlashTool()
-            >>> fw_ver, bl_ver = ft.read_fw_bl_ver()
-            >>> print(f"Firmware: {fw_ver}, Bootloader: {bl_ver}")
-            Firmware: 00010007, Bootloader: 00010002
-
-        Note:
-            - Each version is represented as 4-byte value in big-endian format
-            - The function converts individual bytes to concatenated hexadecimal string
-            - Requires device to be in bootloader mode
-        """
-        tx_row = bytes.fromhex(generate_hex_line(
-            address=0x0000,
-            command=UartBootloaderCmd.UART_COMMAND_READ_PROGRAM_BOOTLOADER_VER,
-            data=[0x00]*8
-        )[1:])
-        logger.debug(f"Sent BiSS Data: {tx_row.hex()}")
-        self._write_to_port(tx_row)
-        response = self.port_read(len(tx_row) - 5)
-        fw_ver = f"{response[0]:02X}{response[1]:02X}{response[2]:02X}{response[3]:02X}"
-        bl_ver = f"{response[4]:02X}{response[5]:02X}{response[6]:02X}{response[7]:02X}"
-        logger.info(f"Firmware version: {fw_ver}, Bootloader version: {bl_ver}")
-        return fw_ver, bl_ver
-
-    def read_memory_state_bl(self) -> UartBootloaderMemoryStates:
-        """
-        Read Memory State of FlashTool bootloader.
-
-        Sends a Read Memory State command to FlashTool.
-
-        Returns:
-            UartBootloaderMemoryStates
-
-        Example:
-            >>> ft = FlashTool()
-            >>> ft.read_memory_state_bl()
-        """
-        tx_row = bytes.fromhex(generate_hex_line(
-            address=0x0000,
-            command=UartBootloaderCmd.UART_COMMAND_READ_MEMORYSTATE,
-            data=[0x00]
-        )[1:])
-        logger.debug(f"Sent BiSS Data: {tx_row.hex()}")
-        self._write_to_port(tx_row)
-        response = self.port_read(len(tx_row) - 5)
-        state = self._decode_memory_state_bl(response)
-        time.sleep(0.01)
-        return state
-
-    def _decode_memory_state_bl(self, response: np.ndarray) -> UartBootloaderMemoryStates:
-        """
-        Decode Memory State of FlashTool bootloader.
-
-        Input:
-            response: value from read_memory_state_bl
-
-        Returns:
-            UartBootloaderMemoryStates
-
-        Example:
-            >>> ft = FlashTool()
-            >>> ft._decode_memory_state_bl(response)
-            UartBootloaderMemoryStates
-        """
-        response = response[0]
-        if response in {state.value for state in UartBootloaderMemoryStates}:
-            matched_state = UartBootloaderMemoryStates(response)
-            logger.debug(f"Response: {matched_state.name} (0x{response:02x})")
-            if matched_state == UartBootloaderMemoryStates.UART_MEMORYSTATE_FLASH_FW_CRC_OK:
-                logger.debug('Firmware CRC check passed!')
-            elif matched_state == UartBootloaderMemoryStates.UART_MEMORYSTATE_FLASH_FW_CRC_FAULT:
-                logger.error('Firmware CRC check failed!')
-            elif matched_state == UartBootloaderMemoryStates.UART_MEMORYSTATE_IDLE:
-                logger.debug('Uart state is IDLE!')
-            elif matched_state == UartBootloaderMemoryStates.UART_MEMORYSTATE_FW_CHECK_CRC32_FAULT:
-                logger.error('Firmware CRC check failed!')
-            elif matched_state == UartBootloaderMemoryStates.UART_MEMORYSTATE_FW_CHECK_CRC32_OK:
-                logger.debug('Firmware CRC check passed!')
-        return matched_state
-
-    def check_main_fw_crc32(self) -> None:
-        """
-        Compare calculated main FlashTool firmware crc32 with ProgramCRC32 value in flash.
-
-        Input:
-            None
-
-        Returns:
-            None
-
-        Example:
-            >>> ft = FlashTool()
-            >>> ft.check_main_fw_crc32()
-            'Response: UART_MEMORYSTATE_FW_CHECK_CRC32_OK (0x05)'
-        """
-        logger.info('Sending check main firmware CRC32 command to FlashTool')
-        tx_row = bytes.fromhex(generate_hex_line(
-            address=0x0000,
-            command=UartBootloaderCmd.UART_COMMAND_CHECK_PROGRAM_CRC32,
-            data=[0x00]
-        )[1:])
-        logger.debug(f"Sent BiSS Data: {tx_row.hex()}")
-        self._write_to_port(tx_row)
-        self.read_memory_state_bl()
-
-    def download_fw_to_ft(self, hex_file_path: str, max_retries: int = 3, pbar: Optional[Any] = None):
-        """
-        Download main firmware from HEX file to FlashTool device using bootloader protocol.
-
-        Implements a robust flash programming routine with:
-        - Automatic CRC verification
-        - Retry mechanism for failed pages
-        - Progress tracking
-        - Error recovery
-
-        Protocol Flow:
-        1. For each 2KB block in HEX file:
-        1.a. Send block's CRC32 checksum first
-        1.b. Transfer data in 64-byte chunks
-        1.c. Verify flash operation success
-        1.d. Retry on failure (up to max_retries)
-        2. Handle both successful and error cases gracefully
-
-        Args:
-            hex_file_path (str): Path to Intel HEX format firmware file
-            max_retries (int): Maximum retry attempts per page (default: 3)
-
-        Raises:
-            RuntimeError: When exceeding max retries for a page
-            IOError: For file access problems
-            ValueError: For invalid HEX file format
-            Exception: For communication errors with device
-
-        Returns:
-            None: Success is implied by normal completion
-
-        Side Effects:
-            - Programs firmware to target device flash
-            - Modifies device memory state
-            - May reset communication interface
-
-        Example:
-            >>> tool = FlashTool()
-            >>> tool.download_fw_to_ft("firmware_v1.2.hex")
-
-        Notes:
-            - Requires active bootloader connection
-            - Uses 2KB page size (device-specific)
-            - 64-byte chunk size optimized for UART throughput
-            - Includes mandatory 1s delay after programming
-            - CRC verification is mandatory for each page
-        """
-        extractor = HexBlockExtractor()
-        retry_count = 0
-        total_pages = 12  # Default value
-        metadata_total_pages = 1
-
-        try:
-            first_block_start, first_block_data, _ = next(extractor.process_hex_file(hex_file_path))
-            first_lower_addr = first_block_start & 0xFFFF
-            first_page_number = first_lower_addr // 2048
-
-            # Extract ProgramLen from metadata (offset 0x0C, 4 bytes)
-            if len(first_block_data) >= 0x0C + 4:
-                program_len_bytes = first_block_data[0x0C:0x0C+4]
-                program_total_pages = int.from_bytes(program_len_bytes, byteorder='little')
-                logger.info(f"Extracted program total pages from metadata: {program_total_pages}")
-
-            # Extract BootloaderLen from metadata (offset 0x1C, 4 bytes)
-            if len(first_block_data) >= 0x1C + 4:
-                program_len_bytes = first_block_data[0x1C:0x1C+4]
-                bootloader_total_pages = int.from_bytes(program_len_bytes, byteorder='little')
-                logger.info(f"Extracted bootloader total pages from metadata: {bootloader_total_pages}")
-
-            total_pages = program_total_pages + bootloader_total_pages + metadata_total_pages
-            logger.info(f"Extracted total pages from metadata: {total_pages}")
-
-            for block_start, block_data, block_crc in extractor.process_hex_file(hex_file_path):
-                lower_addr = block_start & 0xFFFF
-                page_number = lower_addr // 2048
-                count_pages = page_number - first_page_number + 1
-                success = False
-
-                while not success and retry_count < max_retries:
-                    try:
-                        # 1. Send CRC Record
-                        crc_bytes = [
-                            (block_crc >> 24) & 0xFF,
-                            (block_crc >> 16) & 0xFF,
-                            (block_crc >> 8) & 0xFF,
-                            block_crc & 0xFF
-                        ]
-                        crc_line = generate_hex_line(
-                            address=0x0000,
-                            command=UartBootloaderCmd.UART_COMMAND_WRITE_CURRENT_PAGE_CRC32,
-                            data=crc_bytes
-                        )
-                        # logger.debug(f"Sent BiSS Data: {crc_line}")
-                        self.hex_line_send(crc_line)
-
-                        # 2. Send Data Records in 64-byte chunks
-                        for offset in range(0, len(block_data), 64):
-                            chunk = block_data[offset:offset+64]
-                            chunk_addr = (0x01 << 8) | ((page_number) & 0xFF)
-                            data_line = generate_hex_line(
-                                address=chunk_addr,
-                                command=UartBootloaderCmd.UART_COMMAND_LOAD_2K,
-                                data=list(chunk)
-                            )
-                            self.hex_line_send(data_line)
-
-                        if pbar:
-                            percent_complete(count_pages, total_pages, title=f"Sending Page {count_pages}")
-                        # 3. Wait for flash 2048 bytes operation to complete
-                        time.sleep(1)
-
-                        # 4. Verify CRC32 check
-                        matched_state = self.read_memory_state_bl()
-                        if matched_state == UartBootloaderMemoryStates.UART_MEMORYSTATE_FLASH_FW_CRC_FAULT:
-                            retry_count += 1
-                            logger.error(f"CRC Error on page {count_pages}, retry {retry_count}/{max_retries}")
-                            if retry_count >= max_retries:
-                                raise RuntimeError(f"Max retries ({max_retries}) exceeded for page {count_pages}")
-                            continue
-                        elif matched_state == UartBootloaderMemoryStates.UART_MEMORYSTATE_FLASH_FW_CRC_OK:
-                            success = True
-                            retry_count = 0
-                        else:
-                            raise RuntimeError(f"Unexpected memory state: {matched_state.name}")
-
-                    except Exception as e:
-                        retry_count += 1
-                        logger.error(f"Error on page {count_pages}, retry {retry_count}/{max_retries}: {str(e)}")
-                        if retry_count >= max_retries:
-                            raise RuntimeError(f"Max retries ({max_retries}) exceeded for page {count_pages}")
-                        time.sleep(1)
-                        continue
-
-        except Exception as e:
-            print(f"Fatal error during firmware upload: {str(e)}")
-            raise
-        print(end="\n")
-        self.check_main_fw_crc32()
-        time.sleep(0.05)
-
-    def select_spi_channel(self, channel: Literal["channel1", "channel2"]) -> None:
-        """
-        Select SPI communication channel.
-
-        Sends a SELECT CHANNEL command to the FlashTool.
-
-        There are two channels:
-            channel1;
-            channel2.
-
-        Args:
-            channel: "channel1" or "channel2".
-
-        Returns:
-            None
-
-        Raises:
-            ValueError: If the mode is not channel1 or channel2.
-
-        Example:
-            >>> ft = FlashTool()
-            >>> ft.select_SPI_channel('channel1')
-        """
-        channel_mapping = {
-            "channel1": (0, "CHANNEL 1"),
-            "channel2": (1, "CHANNEL 2")
-        }
-
-        if channel not in channel_mapping:
-            raise ValueError(f'Invalid channel: "{channel}". Must be "channel1" or "channel2".')
-
-        channel_num, channel_desc = channel_mapping[channel]
-        logger.info(f"Selected Channel: {channel_num} - {channel_desc}")
-
-        tx_row = bytes.fromhex(generate_hex_line(0, UartCmd.CMD_SELECT_SPI_CH, [channel_num])[1:])
-        self.__port.reset_output_buffer()
-        self.__port.write(tx_row)
-        self.__port.flush()
-
-    def select_flashtool_mode(self, mode: Literal["spi_spi", "ab_uart", "spi_uart_irs", "ab_spi", "default_spi"]) -> None:
-        """
-        Select FlashTool communication mode.
-
-        Sends a SELECT FLASHTOOL MODE command to configure the communication protocol
-        for both channels of the FlashTool.
-
-        Available modes:
-            "spi_spi"      - Channel 1: SPI, Channel 2: SPI
-            "ab_uart"      - Channel 1: AB signal, Channel 2: UART
-            "spi_uart_irs" - Channel 1: SPI, Channel 2: UART for IRS encoders
-            "ab_spi"       - Channel 1: AB signal, Channel 2: SPI
-            "default_spi"  - Default mode: Channel 1: None, Channel 2: SPI
-
-        Args:
-            mode: Communication mode as descriptive string:
-                - "spi_spi"      (0: BISS_MODE_SPI_SPI)
-                - "ab_uart"      (1: BISS_MODE_AB_UART)
-                - "spi_uart_irs" (2: BISS_MODE_SPI_UART_IRS)
-                - "ab_spi"       (3: BISS_MODE_AB_SPI)
-                - "default_spi"  (4: BISS_MODE_DEFAULT_SPI)
-
-        Returns:
-            None
-
-        Raises:
-            ValueError: If invalid mode string provided
-
-        Example:
-            >>> ft = FlashTool()
-            >>> ft.select_flashtool_mode("spi_spi")  # Sets SPI on both channels
-            >>> ft.select_flashtool_mode("spi_uart_irs")  # Sets SPI + UART for IRS
-        """
-        mode_mapping = {
-            "spi_spi": (0, "BISS_MODE_SPI_SPI"),
-            "ab_uart": (1, "BISS_MODE_AB_UART"),
-            "spi_uart_irs": (2, "BISS_MODE_SPI_UART_IRS"),
-            "ab_spi": (3, "BISS_MODE_AB_SPI"),
-            "default_spi": (4, "BISS_MODE_DEFAULT_SPI")
-        }
-
-        if mode not in mode_mapping:
-            valid_modes = list(mode_mapping.keys())
-            raise ValueError(f'Invalid mode: "{mode}". Must be one of: {valid_modes}')
-
-        mode_num, mode_desc = mode_mapping[mode]
-        logger.info(f"Selected FlashTool mode: {mode_num} - {mode_desc}")
-
-        tx_row = bytes.fromhex(generate_hex_line(0, UartCmd.CMD_SELECT_FLASHTOOL_MODE, [mode_num])[1:])
-        self.__port.reset_output_buffer()
-        self.__port.write(tx_row)
-        self.__port.flush()
-        time.sleep(0.05)
-
-    def select_FlashTool_current_sensor_mode(self, mode: Literal["disable", "enable"]) -> None:
-        """
-        Select FlashTool Current sensor mode.
-
-        Sends a SELECT Current sensor mode command to the FlashTool.
-
-        Available modes:
-            "disable" - CURRENT_SENSOR_MODE_DISABLE (0)
-            "enable"  - CURRENT_SENSOR_MODE_ENABLE (1)
-
-        Args:
-            mode: Current sensor mode as descriptive string:
-                - "disable" (0: CURRENT_SENSOR_MODE_DISABLE)
-                - "enable"  (1: CURRENT_SENSOR_MODE_ENABLE)
-
-        Returns:
-            None
-
-        Raises:
-            ValueError: If invalid mode string provided
-
-        Example:
-            >>> ft = FlashTool()
-            >>> ft.select_FlashTool_current_sensor_mode("disable")
-            >>> ft.select_FlashTool_current_sensor_mode("enable")
-        """
-        mode_mapping = {
-            "disable": (0, "CURRENT_SENSOR_MODE_DISABLE"),
-            "enable": (1, "CURRENT_SENSOR_MODE_ENABLE")
-        }
-
-        if mode not in mode_mapping:
-            valid_modes = list(mode_mapping.keys())
-            raise ValueError(f'Invalid mode: "{mode}". Must be one of: {valid_modes}')
-
-        mode_num, mode_desc = mode_mapping[mode]
-        logger.info(f"Selected current sensor mode: {mode_num} - {mode_desc}")
-
-        tx_row = bytes.fromhex(generate_hex_line(0, UartCmd.CMD_SELECT_FLASHTOOL_CURRENT_SENSOR_MODE, [mode_num])[1:])
-        logger.debug(f"Current sensor mode command: {tx_row.hex()}")
-        self.__port.reset_output_buffer()
-        self.__port.write(tx_row)
-        self.__port.flush()
-
-    def select_spi_ch1_mode(self, mode: Literal["lenz_biss", "lir_ssi", "lir_biss_21b"]) -> None:
-        """
-        Select SPI channel 1 mode.
-
-        Sends a SELECT channel 1 SPI mode command to the FlashTool.
-
-        Available modes:
-            "lenz_biss"    - CH1_LENZ_BISS (0)
-            "lir_ssi"      - CH1_LIR_SSI (1)
-            "lir_biss_21b" - CH1_LIR_BISS_21B (2)
-
-        Args:
-            mode: SPI channel 1 mode as descriptive string:
-                - "lenz_biss"    (0: CH1_LENZ_BISS)
-                - "lir_ssi"      (1: CH1_LIR_SSI)
-                - "lir_biss_21b" (2: CH1_LIR_BISS_21B)
-
-        Returns:
-            None
-
-        Raises:
-            ValueError: If invalid mode string provided
-
-        Example:
-            >>> ft = FlashTool()
-            >>> ft.select_spi_ch1_mode("lenz_biss")
-            >>> ft.select_spi_ch1_mode("lir_ssi")
-        """
-        mode_mapping = {
-            "lenz_biss": (0, "CH1_LENZ_BISS"),
-            "lir_ssi": (1, "CH1_LIR_SSI"),
-            "lir_biss_21b": (2, "CH1_LIR_BISS_21B")
-        }
-
-        if mode not in mode_mapping:
-            valid_modes = list(mode_mapping.keys())
-            raise ValueError(f'Invalid mode: "{mode}". Must be one of: {valid_modes}')
-
-        mode_num, mode_desc = mode_mapping[mode]
-        logger.info(f"Selected SPI channel 1 Mode: {mode_num} - {mode_desc}")
-
-        tx_row = bytes.fromhex(generate_hex_line(0, UartCmd.CMD_SELECT_CH1_MODE, [mode_num])[1:])
-        # logger.debug(f"ch1 mode: {tx_row.hex()}")
-        self.__port.reset_output_buffer()
-        self.__port.write(tx_row)
-        self.__port.flush()
-
-    def read_data(self, read_time: float) -> np.ndarray:
-        """
-        Reads encoder data via SPI over USB for a specified duration.
-
-        Sends a command to read data, collects incoming packets, validates them with checksums,
-        and extracts encoder values. Prints progress indicators during reading.
-
-        Args:
-            read_time: The duration in seconds to read data.
-
-        Returns:
-            np.ndarray: A NumPy array of int32 values containing the encoder readings.
-
-        Raises:
-            ValueError: If a checksum or command error occurs in the received data.
-
-        Example:
-            >>> ft = FlashTool()
-            >>> data = ft.read_data(2.0)
-            Read USB data for 2.0 seconds: ..... OK!
-            >>> print(data)
-            array([12345, 12346, ...], dtype=int32)
-        """
-        print('Read USB data for ', read_time, ' seconds: ', end='')
-        size = int(read_time * 500)
-        size_l = int(size) % 256
-        size_m = int(size / 256) % 256
-        self.__port.flushInput()
-        start_time = time.time()
-        self._write_to_port(bytearray([1, size_m, size_l, 129, 0, int((126 - size_m - size_l) % 256)]))
-        ok = 1
-        dot_time = start_time + 0.2
-        new_line_time = start_time + 10
-        gu = []
-        cou = []
-        encoder = []
-        while time.time() < start_time + read_time + 1:
-            if self.__port.inWaiting() > 250:
-                rx_data = np.array(list(self.__port.read(245)), 'int32')
-                crc_hex = - sum(rx_data[0:244]) % 256
-                if (rx_data[0] == 240) & (rx_data[3] == 145):
-                    if crc_hex == rx_data[244]:
-                        cou.extend(rx_data[7:244:4])
-                        encoder.extend(rx_data[6:244:4]*65536 + rx_data[5:244:4]*256 + rx_data[4:244:4])
-                    else:
-                        print('HEX CheckSum ERROR!')
-                        ok = 0
-                        break
-                else:
-                    print('HEX Command ERROR!')
-                    ok = 0
-                    break
-                if (rx_data[1] == 0) & (rx_data[2] == 0):
-                    gu = np.array(gu, 'int32')
-                    encoder = np.array(encoder, 'int32')
-                    break
-            else:
-                time.sleep(0.001)
-            if time.time() > new_line_time:
-                new_line_time = new_line_time + 10
-                print('\n', end='')
-
-            elif time.time() > dot_time:
-                dot_time = dot_time + 0.5
-                print('.', end='')
-        if ok:
-            print(' OK!')
-        return encoder
-
-    def read_data_enc1_enc2_SPI(self, read_time: float, status: bool = True) -> Tuple[List[int], List[int]]:
-        """
-        Reads data from two encoders via SPI over USB for a specified duration.
-
-        Sends a command to read data from both encoders, processes incoming packets, validates checksums,
-        and extracts encoder values for both channels. Optionally prints progress indicators.
-
-        Args:
-            read_time: The duration in seconds to read data.
-            status: If True, prints progress indicators (dots and status messages). Defaults to True.
-
-        Returns:
-            Tuple[List[int], List[int]]: Two lists containing the readings from encoder 1 and encoder 2, respectively.
-
-        Raises:
-            ValueError: If a checksum or command error occurs in the received data.
-
-        Example:
-            >>> ft = FlashTool()
-            >>> enc1, enc2 = ft.read_data_enc1_enc2_SPI(2.0)
-            Read USB data for 2.0 seconds: ..... OK!
-            >>> print(enc1[:5], enc2[:5])
-            [12345, 12346, ...], [54321, 54322, ...]
-        """
-        status and print('Read USB data for ', read_time, ' seconds: ', end='')
-        logger.debug('Reading USB encoder 1 and encoder 2 data for %s seconds.', read_time)
-        size = int(read_time * 1000)
-        size_l = int(size) % 256
-        size_m = int(size / 256) % 256
-        b1 = size_l.to_bytes(1, 'big')
-        b2 = size_m.to_bytes(1, 'big')
-        address = b''.join([b2, b1])
-        addr = int.from_bytes(address, 'big')
-        tx_row = bytes.fromhex(generate_hex_line(addr, UartCmd.HEX_READ_ANGLE_TWO_ENC_SPI, [0])[1:])
-        logger.debug("Reading CMD: %s", generate_hex_line(addr, UartCmd.HEX_READ_ANGLE_TWO_ENC_SPI, [0])[1:])
-        self.__port.flushInput()
-        start_time = time.time()
-        self._write_to_port(tx_row)
-        ok = 1
-        dot_time = start_time + 0.2
-        new_line_time = start_time + 10
-        # gu = []
-        cou1 = []
-        cou2 = []
-        # Encoder1_multiturn = list()
-        # Encoder2_multiturn = list()
-        encoder1 = []
-        encoder2 = []
-        while time.time() < start_time + read_time + 1:
-            if self.__port.inWaiting() > 250:
-                rx_data = np.array(list(self.__port.read(245)), 'int32')
-                crc_hex = - sum(rx_data[0:244]) % 256
-                if (rx_data[0] == 240) & (rx_data[3] == 144):
-                    if crc_hex == rx_data[244]:
-                        cou1.extend(rx_data[7:244:8])
-                        encoder1.extend(rx_data[6:244:8]*65536 + rx_data[5:244:8]*256 + rx_data[4:244:8])
-                        # Encoder_multiturn.extend(rx_data[6:244:6])
-                        cou2.extend(rx_data[11:244:8])
-                        encoder2.extend(rx_data[10:244:8]*65536 + rx_data[9:244:8]*256 + rx_data[8:244:8])
-                    else:
-                        print('HEX CheckSum ERROR!')
-                        ok = 0
-                        break
-                else:
-                    print('HEX Command ERROR!')
-                    ok = 0
-                    break
-                if (rx_data[1] == 0) & (rx_data[2] == 0):
-                    # gu = np.array(gu, 'int32')
-                    # Encoder = np.array(Encoder, 'int32')
-                    break
-            else:
-                time.sleep(0.001)
-            if time.time() > new_line_time:
-                new_line_time = new_line_time + 10
-                status and print('\n', end='')
-
-            elif time.time() > dot_time:
-                dot_time = dot_time + 0.5
-                status and print('.', end='')
-        if ok:
-            status and print(' OK!')
-        return encoder1, encoder2
-
-    def read_data_enc1_AB_enc2_SPI(self, read_time: float, status: bool = True) -> tuple[np.ndarray, np.ndarray]:
-        """
-        Read data from Encoder SPI (SIB, IRS) and AB over USB for a specified duration.
-
-        Sends a command to read data from both encoders, processes incoming packets, validates checksums,
-        and extracts encoder values for both channels. Optionally prints progress indicators.
-
-        Data frame: 6 bytes
-        0 byte  1 byte  2 byte  3 byte  4 byte   5 byte
-        Enc2    *256    *65536  Cou     Enc1     *256
-
-        Returns:
-            tuple[np.ndarray, np.ndarray]: Two np.ndarrays containing:
-                - Encoder2 data.
-                - Encoder1 data.
-        Raises:
-            ValueError: If a checksum or command error occurs in the received data.
-        """
-        status and print('Read USB data for ', read_time, ' seconds: ', end='')
-        size = int(read_time * 750)
-        size_l = int(size) % 256
-        size_m = int(size / 256) % 256
-        b1 = size_l.to_bytes(1, 'big')
-        b2 = size_m.to_bytes(1, 'big')
-        address = b''.join([b2, b1])
-        addr = int.from_bytes(address, 'big')
-        tx_row = bytes.fromhex(generate_hex_line(addr, UartCmd.HEX_READ_ANGLE_TWO_ENC_AB_SPI, [0])[1:])
-
-        self._write_to_port(tx_row)
-
-        Encoder2 = []
-        Encoder1 = []
-        # Cou = list()
-        ok = 1
-        start_time = time.time()
-        dot_time = start_time + 0.2
-        new_line_time = start_time + 10
-
-        try:
-            while time.time() < start_time + read_time + 1:
-                if self.__port.inWaiting() > 250:
-                    rx_data = np.array(list(self.__port.read(245)), 'int32')
-                    CRC_HEX = - sum(rx_data[0:244]) % 256
-                    if (rx_data[0] == 240) & (rx_data[3] == UartCmd.HEX_READ_ANGLE_TWO_ENC_AB_SPI + 0x10):
-                        if (CRC_HEX == rx_data[244]):
-                            # Cou.extend(rx_data[7:244:6])
-                            Encoder2.extend(rx_data[6:244:6]*65536 + rx_data[5:244:6]*256 + rx_data[4:244:6])
-                            Encoder1.extend(rx_data[9:244:6]*256 + rx_data[8:244:6])
-                        else:
-                            ok = 0
-                            raise ValueError("HEX Command ERROR!")
-                    else:
-                        ok = 0
-                        raise ValueError("HEX CheckSum ERROR!")
-
-                    if (rx_data[1] == 0) & (rx_data[2] == 0):
-                        raise ValueError("Recieved Data Size ERROR!")
-                else:
-                    time.sleep(0.001)
-
-                if time.time() > new_line_time:
-                    new_line_time = new_line_time + 10
-                    status and print('\n', end='')
-                elif time.time() > dot_time:
-                    dot_time = dot_time + 0.5
-                    status and print('.', end='')
-            if ok:
-                status and print(' OK!')
-            return np.array(Encoder2, dtype='int32'), np.array(Encoder1, dtype='int32')
-
-        except ValueError as e:
-            logger.error(str(e))
-            return np.array([], dtype='int32'), np.array([], dtype='int32')
-
-    def biss_read_snum(self) -> Optional[Tuple[str, str, str, str]]:
-        """
-        Reads the serial number and metadata from the BiSS encoder.
-
-        Sends a read command to retrieve encoder metadata, including bootloader version,
-        serial number, manufacturing date, and program version. Parses and logs the data, returning it as a tuple.
-
-        Returns:
-            Optional[Tuple[str, str, str, str]]: A tuple containing the bootloader version, serial number,
-            manufacturing date, and program version as hex strings, or None if the read fails.
-
-        Raises:
-            FlashToolError: If the read operation times out or a communication error occurs.
-
-        Example:
-            >>> ft = FlashTool()
-            >>> data = ft.biss_read_snum()
-            >>> print(data)
-            ('0100', '12345678', '20230101', '0200')
-        """
-        try:
-            self._write_to_port(generate_byte_line(BiSSBank.FIXED_ADDRESSES_START_INDEX, UartCmd.HEX_READ_CMD, list(range(64))))
-            self.__port.reset_input_buffer()
-            time.sleep(0.1)
-            cor = BiSSBank.FIXED_ADDRESSES_START_INDEX
-            enc_ver_answ_uint8 = self.port_read(BiSSBank.FIXED_BANK_SIZE)
-            ans = bytes_to_hex_str(enc_ver_answ_uint8)
-            # *2 in indexes is for bytes,
-            endict = {'Bootloader':  ans[(BiSSBank.BOOTLOADER_VER_REG_INDEX-cor)*2:
-                                         (BiSSBank.BOOTLOADER_VER_REG_INDEX-cor+BiSSBank.BOOTLOADER_VER_SIZE)*2],
-                      'Serial No ':  ans[(BiSSBank.DEV_SN_REG_INDEX-cor)*2:
-                                         (BiSSBank.DEV_SN_REG_INDEX-cor+BiSSBank.DEV_SN_SIZE)*2],
-                      'Mfg. Date ':  ans[(BiSSBank.MFG_REG_INDEX-cor)*2:
-                                         (BiSSBank.MFG_REG_INDEX-cor+BiSSBank.MFG_REG_SIZE)*2],
-                      'Program   ':  ans[(BiSSBank.PROGVER_REG_INDEX-cor)*2:
-                                         (BiSSBank.PROGVER_REG_INDEX-cor+BiSSBank.PROGVER_REG_SIZE)*2],
-                      'Dev ID_H  ':  ans[(BiSSBank.DEV_ID_H_REG_INDEX-cor)*2:
-                                         (BiSSBank.DEV_ID_H_REG_INDEX-cor+BiSSBank.DEV_ID_H_SIZE)*2],
-                      'Dev ID_L  ':  ans[(BiSSBank.DEV_ID_L_REG_INDEX-cor)*2:
-                                         (BiSSBank.DEV_ID_L_REG_INDEX-cor+BiSSBank.DEV_ID_L_SIZE)*2]}
-            logger.info('======= ENCODER DATA ========')
-            for name, val in endict.items():
-                logger.info(f'{str(name)}: \t {str(val)}')
-            logger.info('=============================')
-            try:
-                logger.info(f"DEVID: {bytes.fromhex(endict['Dev ID_H  '] + endict['Dev ID_L  ']).decode('ascii')}, " +
-                            f"Serial No: {bytes.fromhex(endict['Serial No '][0:4]).decode('ascii')}" +
-                            f"{endict['Serial No '][4:8]}, " +
-                            f"Mfg date: {time.strftime('%Y-%m-%d %H:%M:%S', time.gmtime(int(endict['Mfg. Date '], 16)))} "
-                            + "(UTC)")
-            except UnicodeDecodeError:
-                pass
-            program_val = endict['Program   ']
-            program_bytes = [f"{int(program_val[i:i+2], 16):02d}" for i in range(0, 8, 2)]
-            program_bytes_reversed = program_bytes[::-1]
-            program_formatted = '.'.join(program_bytes_reversed)
-            bootloader_val = endict['Bootloader']
-            bootloader_bytes = [f"{int(bootloader_val[i:i+2], 16):02d}" for i in range(0, 8, 2)]
-            bootloader_bytes_reversed = bootloader_bytes[::-1]
-            bootloader_formatted = '.'.join(bootloader_bytes_reversed)
-            logger.info(f"Program: {program_formatted}, " +
-                        f"Bootloader: {bootloader_formatted}")
-            logger.debug('Raw encoder answer: %s', ans)
-            return (
-                endict["Bootloader"],
-                endict["Serial No "],
-                endict["Mfg. Date "],
-                endict["Program   "],
-            )
-        except FlashToolError as e:
-            logger.error("ERROR: Can't read registers data! %s", e)
-            return None
-
-    def biss_read_HSI(self) -> Optional[Tuple[str]]:
-        """
-        Reads the HSI (Harmonic Signal Indicator) data from the BiSS encoder.
-
-        Sends a read command to retrieve encoder metadata, including the HSI value, and logs the result.
-        Returns the HSI value as a single-element tuple.
-
-        Returns:
-            Optional[Tuple[str]]: A tuple containing the HSI value as a string, or None if the read fails.
-
-        Raises:
-            FlashToolError: If the read operation times out or a communication error occurs.
-
-        Example:
-            >>> ft = FlashTool()
-            >>> hsi = ft.biss_read_HSI()
-            >>> print(hsi)
-            ('1A',)
-        """
-        try:
-            self._write_to_port(generate_byte_line(BiSSBank.FIXED_ADDRESSES_START_INDEX, UartCmd.HEX_READ_CMD, list(range(64))))
-            time.sleep(0.1)
-            enc_ver_answ_uint8 = self.port_read(BiSSBank.FIXED_BANK_SIZE)
-            enc_ver_answ = bytes_to_hex_str(enc_ver_answ_uint8[4:])
-            enc_ver_dict = {'HSI':  enc_ver_answ_uint8[BiSSBank.FIRSTHARMAMP_REG_INDEX-BiSSBank.FIXED_ADDRESSES_START_INDEX]}
-            logger.info('======= ENCODER DATA ========')
-            logger.debug('Raw encoder answer: %s', enc_ver_answ)
-            for name, val in enc_ver_dict.items():
-                logger.info(f'{str(name)}: \t {str(val)}')
-            logger.info('=============================')
-            return (
-                enc_ver_dict["HSI"]
-            )
-        except FlashToolError as e:
-            logger.error("ERROR: Can't read registers data! %s", e)
-            return None
-
-    def biss_read_progver(self) -> None:
-        """
-        Reads and logs the encoder's program version.
-
-        Retrieves the program version from the BiSS encoder and logs it in a formatted string (e.g., "1.1.0.4").
-
-        Returns:
-            None
-
-        Example:
-            >>> ft = FlashTool()
-            >>> ft.biss_read_progver()
-            INFO: Encoder's program version: 1.1.0.4
-        """
-        logger.info("Encoder's program version: " + ".".join(
-                    f"{num:X}" for num in self.biss_addr_read(BiSSBank.PROGVER_REG_INDEX, 4)[::-1]))
-
-    def biss_read_calibration_temp_vcc(self) -> None:
-        """
-        Continuously reads and prints encoder calibration state, signal modulation, temperature, and VCC.
-
-        Retrieves calibration data, including calibration state, signal modulation, temperature,
-        and supply voltage, and prints them in a loop with a 1-second interval.
-
-        Returns:
-            None
-
-        Example:
-            >>> ft = FlashTool()
-            >>> ft.biss_read_calibration_temp_vcc()
-            CalState: 32, SignalMod: [3671, 4362], EncTemp = 27 °C, Vcc = 4.98 V
-            ...
-        """
-        degree_sign = "\N{DEGREE SIGN}"
-        while True:
-            read_data = self.biss_addr_read(BiSSBank.ENC_DATA_REG_INDEX, 18).view('uint16').byteswap()
-            print(f"CalState: {read_data[0]}, SignalMod: {read_data[[7, 8]]}, ",
-                  f"EncTemp = {int(read_data[1] >> 8) - 64} {degree_sign}C, Vcc = {read_data[2] / 1000} V")
-            time.sleep(1)
-
-    def biss_read_command_state(self) -> Optional[np.ndarray]:
-        """
-        Reads the command state from the BiSS encoder.
-
-        Sends a read command to retrieve the command state,
-        validates the response with a checksum check, and returns the state as a NumPy array.
-
-        Returns:
-            Optional[np.ndarray]: A NumPy array of uint8 values containing the command state, or None if the read fails.
-
-        Raises:
-            FlashToolError: If the checksum validation fails or no data is received within the timeout.
-
-        Example:
-            >>> ft = FlashTool()
-            >>> state = ft.biss_read_command_state()
-            >>> print(state)
-            array([0x01], dtype=uint8)
-        """
-        try:
-            self._write_to_port(generate_byte_line(BiSSBank.CMD_STATE_FLAG_REG_INDEX, UartCmd.HEX_READ_CMD, [1]))
-            time.sleep(0.01)
-            if self.__port.in_waiting >= 1:
-                biss_data = self.__port.read(6)
-                biss_value = int.from_bytes(biss_data, byteorder='big', signed=False)
-                if (calculate_checksum(biss_data[0:-1].hex())) == (biss_data[-1]):
-                    crc_res = "OK"
-                else:
-                    crc_res = "FALSE"
-                    logger.error(f"Received BiSS Data: {biss_value:#010x}, \
-                                checksum calculated {calculate_checksum(biss_data[0:-1].hex())}, \
-                                in data {(biss_data[-1])}, \tres = {crc_res}")
-                    # raise FlashToolError('checksum validation failed for command state.')
-                logger.debug(np.array(list(biss_data[:-1]), 'uint8'))
-                return (np.array(list(biss_data[4:-1]), 'uint8'))
-        except FlashToolError as e:
-            logger.error("ERROR: Can't read command state! %s", e)
-            return None
-
-    def biss_addr_readb(self, bissbank: int, addr: int, length: int) -> np.ndarray:
-        """
-        Reads a specific range of registers from the specified BiSS bank.
-
-        Selects the specified bank, sends a read command for the given address and length,
-        and returns the data as a NumPy array after checksum validation.
-
-        Args:
-            bissbank: The BiSS bank number to read from (0-255).
-            addr: The starting BiSS register address (0-127).
-            length: The number of registers to read.
-
-        Returns:
-            np.ndarray: A NumPy array of uint8 values containing the register data.
-
-        Raises:
-            FlashToolError: If the read operation times out or a checksum error occurs.
-            ValueError: If the address or bank number is out of range.
-
-        Example:
-            >>> ft = FlashTool()
-            >>> data = ft.biss_addr_readb(1, 0x10, 4)
-            >>> print(data)
-            array([0x01, 0x02, 0x03, 0x04], dtype=uint8)
-        """
-        self.biss_set_bank(bissbank)
-        self.__port.flushInput()
-        self._write_to_port(generate_byte_line(addr, UartCmd.HEX_READ_CMD, list(range(length))))
-        time.sleep(0.01)
-
-        if self._wait_for_data(length + 1, timeout=1.0):
-            biss_data = self.__port.read(length + 5)  # len, addr, addr, cmd, crc
-            biss_value = int.from_bytes(biss_data, byteorder='big', signed=False)
-            calculated_crc = calculate_checksum(biss_data[0:-1].hex())
-            if calculated_crc == biss_data[-1]:
-                crc_res = "OK"
-                logger.debug(f"Received BiSS Data: {biss_value:#010x}, checksum calculated {calculated_crc}, \
-                             in data {biss_data[-1]}, res = {crc_res}")
-            else:
-                crc_res = "FALSE"
-                logger.error(f"Received BiSS Data: {biss_value:#010x}, checksum calculated {calculated_crc}, \
-                             in data {biss_data[-1]}, res = {crc_res}")
-            logger.debug("BiSS Bank %s registers at %s:", bissbank, addr)
-            logger.debug(np.array(list(biss_data[4:-1]), 'uint8'))
-            return np.array(list(biss_data[4:-1]), 'uint8')
-            # logger.info(np.array(list(biss_data[0:64]), 'uint8'))
-            # logger.info(np.array(list(biss_data[64:-1]), 'uint8'))
-        raise FlashToolError('Timeout waiting for register data.')
-
-    def biss_addr_read(self, addr: int, length: int) -> np.ndarray:
-        """
-        Reads a specific range of registers from the current BiSS bank.
-
-        Sends a read command for the given address and length, and returns the data as a NumPy array after checksum validation.
-
-        Args:
-            addr: The starting BiSS register address (0-127).
-            length: The number of registers to read.
-
-        Returns:
-            np.ndarray: A NumPy array of uint8 values containing the register data.
-
-        Raises:
-            FlashToolError: If the read operation times out or a checksum error occurs.
-            ValueError: If the address is out of range.
-
-        Example:
-            >>> ft = FlashTool()
-            >>> data = ft.biss_addr_read(0x10, 4)
-            >>> print(data)
-            array([0x01, 0x02, 0x03, 0x04], dtype=uint8)
-        """
-        self.__port.flushInput()
-        self._write_to_port(generate_byte_line(addr, UartCmd.HEX_READ_CMD, list(range(length))))
-        time.sleep(0.01)
-
-        if self._wait_for_data(length + 1, timeout=1.0):
-            biss_data = self.__port.read(length + 5)  # len, addr, addr, cmd, crc
-            biss_value = int.from_bytes(biss_data, byteorder='big', signed=False)
-            calculated_crc = calculate_checksum(biss_data[0:-1].hex())
-            if calculated_crc == biss_data[-1]:
-                crc_res = "OK"
-                # logger.debug(f"Received BiSS Data: {biss_value:#010x}, checksum calculated {calculated_crc}, \
-                #              in data {biss_data[-1]}, res = {crc_res}")
-            else:
-                crc_res = "FALSE"
-                logger.error(f"Received BiSS Data: {biss_value:#010x}, checksum calculated {calculated_crc}, \
-                             in data {biss_data[-1]}, res = {crc_res}")
-                raise FlashToolError('checksum error.')
-            logger.debug("Registers at %s:", addr)
-            logger.debug(np.array(list(biss_data[4:-1]), 'uint8'))
-            return np.array(list(biss_data[4:-1]), 'uint8')
-            # logger.info(np.array(list(biss_data[0:64]), 'uint8'))
-            # logger.info(np.array(list(biss_data[64:-1]), 'uint8'))
-        raise FlashToolError('Timeout waiting for register data.')
-
-    def biss_read_flags_flashCRC(self) -> int:
-        """
-        Reads the flash CRC error flag from the BiSS encoder.
-
-        Retrieves the state flags and extracts the flash CRC error flag (bit 1).
-
-        Returns:
-            int: The flash CRC flag value (0 or 1).
-
-        Raises:
-            FlashToolError: If the read operation fails or CRC validation fails.
-
-        Example:
-            >>> ft = FlashTool()
-            >>> crc_flag = ft.biss_read_flags_flashCRC()
-            >>> print(crc_flag)
-            0
-        """
-        state_flags = self.biss_read_state_flags()
-        error_flags_flash_pos = 1  # TODO use ERROR_FLAGS here
-        crc_flag = state_flags >> error_flags_flash_pos & 1
-        return crc_flag
-
-    def biss_read_flags(self) -> Tuple[list[str], list[str]]:
-        """
-        Reads and interprets state flags and command state from the BiSS encoder.
-
-        Retrieves state flags and command state, interprets them using helper functions,
-        and logs the results. Returns the interpreted error flags and command state descriptions.
-
-        Returns:
-            Tuple[List[str], List[str]]: A tuple containing:
-                - A list of active error flag descriptions.
-                - A list containing the command state description.
-
-        Raises:
-            FlashToolError: If reading state flags or command state fails.
-
-        Example:
-            >>> ft = FlashTool()
-            >>> flags, cmd_state = ft.biss_read_flags()
-            >>> print(flags, cmd_state)
-            ['FLASH_CRC_ERROR'], ['IDLE']
-        """
-        try:
-            state_flags = self.biss_read_state_flags()
-            if state_flags is None:
-                logger.error("Failed to read state flags: No data received")
-                raise FlashToolError("No state flags data received from encoder")
-            logger.debug("State flags raw data: %s", state_flags)
-            flags = (np.uint16(state_flags[1]) << 8) | state_flags[0]
-            interpreted_flags = interpret_error_flags(flags)
-            logger.info("Interpreted error flags: %s", interpreted_flags)
-            command_state = self.biss_read_command_state()[0]
-            logger.debug("Command state raw data: %s", command_state)
-            interpreted_command_state = interpret_biss_commandstate(command_state)
-            logger.info("Interpreted command state: %s", interpreted_command_state)
-
-            return interpreted_flags, interpreted_command_state
-
-        except FlashToolError as e:
-            logger.error("Failed to read flags: %s", e)
-            raise
-
-    def biss_read_angle_once(self) -> None:
-        """
-        Reads the encoder angle once and prints it in degrees, minutes, and seconds.
-
-        Retrieves a single angle reading from encoder 2, converts it to degrees, and prints the formatted output.
-
-        Returns:
-            None
-
-        Example:
-            >>> ft = FlashTool()
-            >>> ft.biss_read_angle_once()
-            [2119664]:    45° 30' 15"
-        """
-        degree_sign = "\N{DEGREE SIGN}"
-        try:
-            # _, ans = self.read_data_enc1_enc2_SPI(0.01, False)
-            ans, _ = self.read_data_enc1_AB_enc2_SPI(0.01, False)
-            if not ans.any() or len(ans) < 1:
-                logger.error("No valid data received from encoder")
-                return
-
-            raw_value = int(ans[0])
-            resolution = 2**24
-            total_degrees = raw_value * 360.0 / resolution
-
-            degrees = int(total_degrees)
-            remaining = total_degrees - degrees
-            minutes = int(remaining * 60)
-            seconds = round((remaining * 60 - minutes) * 60, 2)
-
-            output = (f"[{raw_value}]: \t"
-                      f"{degrees:>3}{degree_sign} "
-                      f"{minutes:02d}' "
-                      f"{seconds:05.2f}\"")
-
-            sys.stdout.write("\r" + output + " " * 10)
-            logger.debug(output)
-
-        except ValueError as e:
-            logger.error(f"Invalid encoder data: {e}")
-        except Exception as e:
-            logger.error(f"Error reading angle: {e}")
-
-        # degree_sign = "\N{DEGREE SIGN}"
-        # _, ans = self.read_data_enc1_enc2_SPI(0.01, False)
-        # res = 2**24
-        # ang = int(ans[0]) * 360 / res
-        # degrs = int(ang)
-        # mins = int((ang - degrs) * 60)
-        # secs = int((ang - degrs - (mins / 60)) * 3600)
-        # sys.stdout.write("\r" + f'[{ans[0]}]: \t {str(degrs):>3}{degree_sign} {str(mins):2}\' {str(secs):2}\"' + '\t\t')
-        # logger.debug(f'[{ans[0]}]: \t {str(degrs):>3}{degree_sign} {str(mins):2}\' {str(secs):2}\"')
-
-    def biss_zeroing(self) -> None:
-        """
-        Performs a zeroing calibration on the BiSS encoder.
-
-        Resets the FlashTool, power cycles the encoder, unlocks setup and flash,
-        issues the zeroing command, saves to flash, and power cycles again.
-
-        Returns:
-            None
-
-        Example:
-            >>> ft = FlashTool()
-            >>> ft.biss_zeroing()
-        """
-        self.flashtool_rst()
-        self.encoder_power_cycle()
-
-        self.biss_write_command('unlocksetup')
-        self.biss_write_command('unlockflash')
-        self.biss_write_command('zeroing')
-        self.biss_write_command('saveflash')
-        time.sleep(0.2)
-        self.encoder_power_cycle()
-
-    def biss_set_dir_cw(self) -> None:
-        """
-        Sets the encoder direction to clockwise.
-
-        Resets the FlashTool, power cycles the encoder, unlocks setup and flash,
-        issues the clockwise direction command, saves to flash, and power cycles again.
-
-        Returns:
-            None
-
-        Example:
-            >>> ft = FlashTool()
-            >>> ft.biss_set_dir_cw()
-        """
-        self.flashtool_rst()
-        self.encoder_power_cycle()
-
-        self.biss_write_command('unlocksetup')
-        self.biss_write_command('unlockflash')
-        self.biss_write_command('set_dir_cw')
-        self.biss_write_command('saveflash')
-        time.sleep(0.2)
-        self.encoder_power_cycle()
-
-    def biss_set_dir_ccw(self) -> None:
-        """
-        Sets the encoder direction to counterclockwise.
-
-        Resets the FlashTool, power cycles the encoder, unlocks setup and flash,
-        issues the counterclockwise direction command, saves to flash, and power cycles again.
-
-        Returns:
-            None
-
-        Example:
-            >>> ft = FlashTool()
-            >>> ft.biss_set_dir_ccw()
-        """
-        self.flashtool_rst()
-        self.encoder_power_cycle()
-
-        self.biss_write_command('unlocksetup')
-        self.biss_write_command('unlockflash')
-        self.biss_write_command('set_dir_ccw')
-        self.biss_write_command('saveflash')
-        time.sleep(0.2)
-        self.encoder_power_cycle()
-
-    def biss_set_shift(self, shift_angle: int) -> None:
-        """
-        Sets the encoder shift angle and verifies the operation.
-
-        Reads the current angle, unlocks setup and flash, writes the shift angle,
-        saves to flash, power cycles the encoder, and verifies the new shift angle.
-
-        Args:
-            shift_angle: The shift angle value to set.
-
-        Returns:
-            None
-
-        Example:
-            >>> ft = FlashTool()
-            >>> ft.biss_set_shift(1000)
-            [12345]:     45° 30' 15" # TODO
-            [1000]:      0°  0'  0"
-        """
-        if not isinstance(shift_angle, int):
-            raise ValueError(f"Shift angle must be an integer, got {type(shift_angle)}")
-        if not 0 <= shift_angle <= 255:  # TODO CHECK
-            raise ValueError(f"Shift angle must be in range [0, {2**24 - 1}], got {shift_angle}")
-
-        self.biss_read_angle_once()
-        self.biss_write_command('unlocksetup')
-        self.biss_write_command('unlockflash')
-        self.biss_write_word(BiSSBank.SHIFT_REG_INDEX, 5, shift_angle)  # TODO add validation
-        self.biss_write_command('saveflash')
-        time.sleep(0.1)
-        print(self.biss_addr_read(BiSSBank.SHIFT_REG_INDEX, 1))
-
-        self.encoder_power_cycle()
-        print(self.biss_addr_read(BiSSBank.SHIFT_REG_INDEX, 1))
-        self.biss_read_angle_once()
-
-    def send_data_to_device(self,
-                            pages: List[List[bytes]],
-                            crc_values: List[int],
-                            page_numbers: List[int],
-                            start_page: int,
-                            end_page: int,
-                            pbar: Optional[Any] = None,
-                            difmode: bool = False
-                            ) -> None:
-        """
-        Transmits organized data pages to the BiSS encoder with CRC verification.
-
-        Sends pages of data to the encoder, writing CRC and page numbers as needed,
-        and verifies each page with a flash CRC check. Supports a progress bar and differential mode for specific use cases.
-
-        Args:
-            pages: A list of pages, where each page is a list of byte arrays.
-            crc_values: A list of CRC checksums for each page.
-            page_numbers: A list of page numbers corresponding to each page.
-            start_page: The index of the first page to transmit.
-            end_page: The index of the last page to transmit.
-            pbar: An optional progress bar object for tracking transmission progress.
-            difmode: If True, uses differential table transmission mode. Defaults to False.
-
-        Raises:
-            SystemExit: If transmission fails after the maximum number of retries.
-            FlashToolError: If a hardware communication error occurs.
-
-        Example:
-            >>> ft = FlashTool()
-            >>> pages = [[b'\\x01\\x02', b'\\x03\\x04'], [b'\\x05\\x06', b'\\x07\\x08']]
-            >>> crc_values = [0x1234, 0x5678]
-            >>> page_numbers = [1, 2]
-            >>> ft.send_data_to_device(pages, crc_values, page_numbers, 0, 1)
-            INFO: Done uploading!
-        """
-        max_retries = 3
-
-        for page_idx, (page_data, crc, page_num) in enumerate(zip(pages, crc_values, page_numbers), start=start_page):
-            if page_idx > end_page:
-                break
-
-            retry_count = 0
-            success = False
-
-            while retry_count < max_retries and not success:
-                # Write CRC and page number for pages after the first
-                if page_idx > 1:
-                    self.biss_set_bank(BiSSBank.BISS_BANK_SERV)
-                    self.biss_write_word(BiSSBank.CRC32_REG_INDEX, crc)
-                    self.biss_write_word(BiSSBank.PAGENUM_REG_INDEX, page_num)
-                    time.sleep(0.01)
-                    # self.biss_read_registers(BISS_BANK_SERV)
-
-                # Send each bank in the page
-                for bank_idx, bank_data in enumerate(page_data):
-                    bank_num = bank_idx + BiSSBank.BISS_USERBANK_START
-                    if bank_num == 5:  # Set bank 5 explicitly
-                        self.biss_set_bank(bank_num)
-                    time.sleep(0.05)
-
-                    if pbar:
-                        percent_complete(bank_idx, BiSSBank.BANKS_PER_PAGE - 1, title=f"Sending Page {page_idx}")
-                    # print('HEX-LINE', generate_hex_line(0, HEX_WRITE_CMD, bank_data))
-                    hex_line = generate_hex_line(0, UartCmd.HEX_WRITE_CMD, bank_data)
-                    self.hex_line_send(hex_line)
-
-                time.sleep(0.3)
-                if difmode:
-                    self.biss_write_command('savediftable')
-                else:
-                    self.biss_write_command('load2k')
-
-                time.sleep(1.25)
-
-                if not self.biss_read_flags_flashCRC()[0]:
-                    success = True
-                    logger.debug("Page %s sent successfully", page_idx)
-                else:
-                    retry_count += 1
-                    logger.error("Page %s CRC mismatch! Retry %s/%s", page_idx, retry_count, max_retries)
-                    # self.power_cycle()
-                    # self.biss_cmd_reboot2bl()
-                    if retry_count == max_retries:
-                        logger.critical("Failed to send page %s after %s attempts", page_idx, max_retries)
-                        sys.exit(1)
-
-        logger.info(" Done uploading!")
-
-    def read_enc2_current(self) -> tuple[str, float] | bool:
-        """
-        Read current of the encoder on channel 2.
-
-        Args:
-            None
-
-        Returns:
-            tuple[str, int]: If successful, returns:
-                - str: Encoder2 current in hexadecimal format
-                - float: Current in mA
-            bool: False if operation fails (CRC error, no response, etc.)
-        """
-        try:
-            self.__port.reset_output_buffer()
-            self.__port.reset_input_buffer()
-            self.__port.write(generate_byte_line(0, UartCmd.HEX_READ_ENC2_CURRENT, list(range(UartCmd.RX_DATA_LENGTH_CURRENT))))
-            self.__port.flush()
-
-            enc_ans = self.__port.read(UartCmd.RX_DATA_LENGTH_CURRENT + UartCmd.PKG_INFO_LENGTH)
-
-            logger.debug(enc_ans.hex())
-
-            if not enc_ans:
-                logger.error("No response from encoder!")
-                return False
-
-            enc_data_np = np.array(list(enc_ans), dtype='uint8')
-
-            if (enc_data_np[0] != enc_data_np.size - UartCmd.PKG_INFO_LENGTH) or \
-               (enc_data_np[3] != UartCmd.HEX_READ_ENC2_CURRENT + UartCmd.CMD_VAL_ADD):
-                logger.error("Invalid response structure from encoder!")
-                return False
-
-            calculated_crc = calculate_checksum(enc_ans[0:-1].hex())
-            if calculated_crc != enc_data_np[-1]:
-                logger.error(f"CRC mismatch: calculated {calculated_crc}, expected {enc_data_np[-1]}")
-                return False
-
-            logger.debug("CRC check passed.")
-
-            enc2_current = enc_ans[4:4+UartCmd.RX_DATA_LENGTH_CURRENT]
-
-            data_hex = enc2_current.hex()
-
-            ans_enc2_current_ma = int.from_bytes(enc2_current, byteorder='little', signed=False) / 1000
-
-            return data_hex, ans_enc2_current_ma
-
-        except Exception as e:
-            logger.error(f"Error reading encoder current: {str(e)}", exc_info=True)
-            return False
-
-    def read_instant_angle_enc_SPI(self) -> tuple[str, list[int]] | bool:
-        """
-        Read instant angle encoder via SPI over USB.
-
-        Returns:
-            tuple[str, list[int]] | bool:
-                If successful, returns a tuple containing:
-                    - str: Encoder angle in hexadecimal format (24-bit value)
-                    - list[int]: Angle parts [degrees, minutes, seconds]
-                bool: False if operation fails
-                    - No response from encoder
-                    - Invalid response structure
-                    - CRC mismatch in packet checksum
-                    - Communication errors
-
-        Example:
-            >>> ft = FlashTool()
-            >>> result = ft.read_instant_angle_enc_SPI()
-            [16733568]:    359° 03' 48"
-        """
-        RX_DATA_LENGTH = 4
-        DEGREE_SIGN = "\N{DEGREE SIGN}"
-
-        try:
-            self.__port.reset_output_buffer()
-            self.__port.reset_input_buffer()
-            self.__port.write(generate_byte_line(
-                address=0x0000,
-                command=UartCmd.HEX_READ_INSTANT_ANGLE_ENC_SPI,
-                data=list(range(RX_DATA_LENGTH))
-            ))
-            self.__port.flush()
-
-            enc_ans = self.__port.read(RX_DATA_LENGTH + UartCmd.PKG_INFO_LENGTH)
-
-            if not enc_ans:
-                logger.error("No response from encoder!")
-                return False
-
-            enc_data_np = np.array(list(enc_ans), dtype='uint8')
-
-            if (enc_data_np[0] != enc_data_np.size - UartCmd.PKG_INFO_LENGTH) or \
-               (enc_data_np[3] != UartCmd.HEX_READ_INSTANT_ANGLE_ENC_SPI + UartCmd.CMD_VAL_ADD):
-                logger.error("Invalid response structure from encoder!")
-                return False
-
-            calculated_crc = calculate_checksum(enc_ans[0:-1].hex())
-            if calculated_crc != enc_data_np[-1]:
-                logger.error(f"CRC mismatch: calculated {calculated_crc}, expected {enc_data_np[-1]}")
-                return False
-
-            logger.debug("CRC check passed.")
-
-            angle_data = enc_ans[4:4+RX_DATA_LENGTH]
-
-            data_hex = angle_data.hex()
-
-            ans_angle = int.from_bytes(angle_data[:3], byteorder='little', signed=False)
-
-            angle_raw = (
-                int(data_hex[4:6], 16) * 65536 +
-                int(data_hex[2:4], 16) * 256 +
-                int(data_hex[0:2], 16)
-            )
-            angle_deg = angle_raw * 360 / 2**24
-
-            degrees = int(angle_deg)
-            remaining = angle_deg - degrees
-            minutes = int(remaining * 60)
-            seconds = int((remaining * 60 - minutes) * 60)
-
-            logger.info(
-                f"[{ans_angle}]: {degrees:>3}{DEGREE_SIGN} "
-                f"{minutes:02d}' {seconds:02d}\""
-            )
-
-            return ans_angle, [degrees, minutes, seconds]
-
-        except Exception as e:
-            logger.error(f"Error reading encoder angle: {str(e)}", exc_info=True)
-            return False
-
-    def read_instant_angle_packet_enc_SPI(self) -> tuple[str, list[int]] | bool:
-        """
-        Read instant extended angle packet from encoder via SPI over USB.
-
-        This method reads not only the angle value but also additional status
-        information including error/warning flags and CRC for data integrity
-        verification.
-
-        Returns:
-            tuple[str, list[int]] | bool:
-                If successful, returns a tuple containing:
-                    - str: Encoder angle in hexadecimal format (24-bit value)
-                    - list[int]: Extended angle information:
-                        [degrees, minutes, seconds, nE_flag, nW_flag, received_crc, status]
-                        where:
-                        - degrees: Integer degrees (0-359)
-                        - minutes: Integer minutes (0-59)
-                        - seconds: Integer seconds (0-59)
-                        - nE_flag: Error flag (0=no error, 1=error detected)
-                        - nW_flag: Warning flag (0=no warning, 1=warning present)
-                        - received_crc: Received 6-bit CRC value from encoder
-                        - status: "OK" if CRC matches, "ERR" if CRC mismatch
-                bool: False if operation fails due to:
-                    - No response from encoder
-                    - Invalid response structure
-                    - CRC mismatch in packet checksum
-                    - Communication errors
-
-        Example:
-            >>> ft = FlashTool()
-            >>> result = ft.read_instant_angle_packet_enc_SPI()
-            [16733568]:    359° 03' 48"
-            nE:1 nW:1 CRC:0F (OK)
-        """
-        RX_DATA_LENGTH = 6
-        DEGREE_SIGN = "\N{DEGREE SIGN}"
-
-        try:
-            self.__port.reset_input_buffer()
-            self.__port.reset_output_buffer()
-            self.__port.write(generate_byte_line(
-                address=0x0000,
-                command=UartCmd.HEX_READ_INSTANT_ANGLE_PACKET_ENC_SPI,
-                data=list(range(RX_DATA_LENGTH))
-            ))
-            self.__port.flush()
-
-            enc_ans = self.__port.read(RX_DATA_LENGTH + UartCmd.PKG_INFO_LENGTH)
-
-            if not enc_ans:
-                logger.error("No response from encoder!")
-                return False
-
-            enc_data_np = np.array(list(enc_ans), dtype='uint8')
-
-            if (enc_data_np[0] != enc_data_np.size - UartCmd.PKG_INFO_LENGTH) or \
-               (enc_data_np[3] != UartCmd.HEX_READ_INSTANT_ANGLE_PACKET_ENC_SPI + UartCmd.CMD_VAL_ADD):
-                logger.error("Invalid response structure from encoder!")
-                return False
-
-            calculated_crc = calculate_checksum(enc_ans[0:-1].hex())
-            if calculated_crc != enc_data_np[-1]:
-                logger.error(f"CRC mismatch: calculated {calculated_crc}, expected {enc_data_np[-1]}")
-                return False
-
-            logger.debug("CRC check passed.")
-
-            angle_data = enc_ans[4:4+RX_DATA_LENGTH]
-
-            angle_bytes = angle_data[:3]
-            nE_nW_bits = angle_data[4] & 0x03
-            nE = (nE_nW_bits >> 1) & 0x01
-            nW = nE_nW_bits & 0x01
-            received_crc = angle_data[5]
-
-            angle_value = int.from_bytes(angle_bytes, byteorder='little', signed=False)
-            data_for_crc = np.uint32(angle_value << 2) | np.uint32(nE_nW_bits)
-            expected_crc = biss_crc6_calc(data_for_crc)
-
-            angle_deg = angle_value * 360 / 2**24
-            degrees = int(angle_deg)
-            remaining = angle_deg - degrees
-            minutes = int(remaining * 60)
-            seconds = int((remaining * 60 - minutes) * 60)
-
-            crc_ok = (expected_crc == received_crc)
-            status = "OK" if crc_ok else "ERR"
-
-            logger.info(
-                f"[{angle_value}]: {degrees:>3}{DEGREE_SIGN} "
-                f"{minutes:02d}' {seconds:02d}\"\n"
-                f"nE:{nE:01X} nW:{nW:01X} CRC:{received_crc:02X} "
-                f"({status})"
-            )
-
-            return angle_value, [degrees, minutes, seconds, nE, nW, received_crc, status]
-
-        except Exception as e:
-            logger.error(f"Error reading encoder angle: {str(e)}", exc_info=True)
-            return False
-
-    def enter_bl_biss_encoder(self, reset_attempts: int = 12, power_cycle_delay: float = 0.01) -> bool:
-        """
-        Reset the BISS encoder to bootloader mode by power cycling.
-
-        This method performs multiple power cycles to force the encoder
-        into bootloader mode. The bootloader mode is indicated by specific
-        error flags being set in the encoder's status register.
-
-        Bootloader Mode Indicators:
-        - The presence of ['FLAGS_STARTUP_ERROR'] flags typically indicates
-        successful entry into bootloader mode.
-
-        Args:
-            reset_attempts (int): Number of power cycle attempts (default: 12)
-            power_cycle_delay (float): Delay between power cycles in seconds (default: 0.01)
-
-        Returns:
-            bool: True if encoder entered bootloader mode (startup error flags detected),
-                False otherwise.
-        """
-        initial_flags, initial_cmd_state = self.biss_read_flags()
-        # logger.debug("Initial flags: %s, Command state: %s", initial_flags, initial_cmd_state)
-
-        for attempt in range(reset_attempts):
-            self.encoder_power_off()
-            time.sleep(power_cycle_delay)
-            self.encoder_power_on()
-            time.sleep(power_cycle_delay)
-
-        final_flags, final_cmd_state = self.biss_read_flags()
-        # logger.debug("Final flags after reset: %s, Command state: %s", final_flags, final_cmd_state)
-
-        bootloader_entered = 'FLAGS_STARTUP_ERROR' in final_flags
-
-        logger.info("Encoder enter bootloader %s", "successful" if bootloader_entered else "failed")
-        return bootloader_entered
-
-    def enter_bl_irs(self) -> bool:
-        """
-        Enter bootloader of IRS encoder and verify successful connection.
-
-        Workflow:
-        - Selects SPI channel 2 for IRS encoder communication
-        - Sets flashtool mode to 'spi_uart_irs' for proper protocol configuration
-        - Powers off the encoder briefly (100ms) then powers it back on
-        - Sends the bootloader stay command sequence
-        - Validates the encoder's response against expected bootloader acknowledgment
-
-        Args:
-            None
-
-        Returns:
-            bool: True if the encoder successfully enters bootloader mode and responds
-              correctly, False otherwise.
-
-        Example:
-            >>> encoder_handler.reboot_to_bl_irs()
-            True  # Encoder successfully entered bootloader mode
-        """
-        try:
-            self.select_spi_channel('channel2')
-            self.select_flashtool_mode('spi_uart_irs')
-            self.encoder_power_off()
-            time.sleep(0.1)
-            self.encoder_power_on()
-            time.sleep(0.01)
-
-            tx_row = bytes.fromhex(generate_hex_line(
-                address=0x0000,
-                command=UartCmd.HEX_IRS_ENC_WRITE_READ_CMD,
-                data=UartBootloaderSeq.UART_SEQ_STAY_IN_BL,
-            )[1:])
-            self._write_to_port(tx_row)
-
-            enc_ans = self.port_read(len(tx_row)-1)
-            if enc_ans is None or len(enc_ans) == 0:
-                logger.error("No response from IRS encoder!")
-                return False
-
-            expected_ans = np.array(UartBootloaderSeq.UART_SEQ_ANSWER_TO_STAY_IN_BL, dtype=np.uint8)
-
-            if np.array_equal(expected_ans, enc_ans):
-                logger.info("IRS Encoder enter bootloader successfully!")
-                return True
-            else:
-                logger.error("Failed to enter IRS bootloader! Unexpected response.")
-                logger.debug(f"Expected: {expected_ans.tobytes().hex()}")
-                logger.debug(f"Received: {enc_ans.tobytes().hex()}")
-                return False
-
-        except Exception as e:
-            logger.error(f"An exception occurred while connecting to encoder: {e}")
-            return False
-
-    def enter_fw_irs(self) -> bool:
-        """
-        Exit bootloader and reboot the IRS encoder into normal firmware operation.
-
-        Workflow:
-        - Selects SPI channel 2 and configures 'spi_uart_irs' communication mode
-        - Cycles encoder power to ensure clean state transition
-        - Sends the bootloader exit command sequence
-        - Validates the encoder's acknowledgment of bootloader exit
-        - Always restores the flashtool to default SPI mode as cleanup
-
-        Returns:
-            bool: True if the encoder successfully exits bootloader mode and responds
-              correctly, False otherwise. The system is always returned to default
-              SPI mode regardless of the result.
-
-        Example:
-            >>> encoder_handler.reboot_to_fw_irs()
-            True  # Encoder successfully exited bootloader and returned to firmware mode
-        """
-        def _cleanup_and_return(success: bool) -> bool:
-            self.select_flashtool_mode('default_spi')
-            return success
-
-        try:
-            self.select_spi_channel('channel2')
-            self.select_flashtool_mode('spi_uart_irs')
-            self.encoder_power_off()
-            time.sleep(0.1)
-            self.encoder_power_on()
-            time.sleep(0.01)
-            tx_row = bytes.fromhex(generate_hex_line(
-                address=0x0000,
-                command=UartCmd.HEX_IRS_ENC_WRITE_READ_CMD,
-                data=UartBootloaderSeq.UART_SEQ_EXIT_BL,
-            )[1:])
-            self._write_to_port(tx_row)
-
-            enc_ans = self.port_read(len(tx_row)-1)
-
-            if enc_ans is None or len(enc_ans) == 0:
-                logger.error("No response from IRS encoder!")
-                return _cleanup_and_return(False)
-
-            expected_ans = np.array(UartBootloaderSeq.UART_SEQ_ANSWER_TO_EXIT_BL, dtype=np.uint8)
-
-            if np.array_equal(expected_ans, enc_ans):
-                logger.info("IRS Encoder disconnect from bootloader successfully!")
-                return _cleanup_and_return(True)
-            else:
-                logger.error("Failed to disconnect bootloader! Unexpected response.")
-                logger.debug(f"Expected: {expected_ans.hex()}")
-                logger.debug(f"Received: {enc_ans.hex()}")
-                return _cleanup_and_return(False)
-
-        except Exception as e:
-            logger.error(f"An exception occurred while disconnecting from bootloader: {e}")
-            return _cleanup_and_return(False)
-
-    def set_pos_irs(self, deg: float, reverse: bool = False) -> tuple:
-        """
-        Set IRS encoder position.
-        
-        Args:
-            deg (float): Desired angle in degrees (0 to 360).
-            reverse (bool): If True, use reverse direction mode.
-        
-        Returns:
-            tuple: (b_1, b_2) or (None, None) on failure.
-        """
-        if not (0 <= deg <= 360):
-            logger.error(f"Invalid angle {deg}. Must be between 0 and 360 degrees.")
-            return None, None
-        
-        # Mode configuration
-        offset = 13 if reverse else 5
-        cmd_byte = 14 if reverse else 6
-        checksum_base = 242 if reverse else 250
-        
-        try:
-            data = np.int32((deg / 360 * 4096) % 4096)
-            b_2 = np.uint8(data >> 4)
-            b_1 = np.uint8(data << 4) | offset
-            
-            if not self.enter_bl_irs():
-                logger.error("Failed to connect to bootloader of IRS encoder!")
-                return None, None
-            
-            set_pos_cmd = bytearray([0, b_1, b_2, cmd_byte, 
-                                    np.uint8(checksum_base - b_1 - b_2)])
-            logger.debug(f"Sending {'reverse' if reverse else 'forward'} command: {set_pos_cmd.hex().upper()}")
-            
-            tx = bytes.fromhex(generate_hex_line(
-                address=0x0000,
-                command=UartCmd.HEX_IRS_ENC_WRITE_READ_CMD,
-                data=set_pos_cmd,
-            )[1:])
-            self._write_to_port(tx)
-            
-            enc_ans = self.port_read(len(tx) - UartCmd.PKG_INFO_LENGTH)
-            
-            if enc_ans is None:
-                logger.error("No response from IRS encoder!")
-                return None, None
-            
-            if not self.enter_fw_irs():
-                logger.error("Failed to exit bootloader mode of IRS encoder!")
-                return None, None
-            
-            return b_1, b_2
-            
-        except Exception as e:
-            logger.error(f"Exception while setting IRS encoder position: {e}")
-            return None, None
+        logger.info('FlashTool: %s closed.', self._port_name)
